==> 01_historical-occurrence-data.Rmd <==
---
editor_options: 
  chunk_output_type: console
---

# Analyzing historical occurrence data

This script carries out exploratory analysis of historical occurrence data from the Nilgiri hills. 
```{r}
# Load necessary libraries
library(dplyr)
library(stringr)
library(tidyverse)
library(scico)
library(RColorBrewer)
library(extrafont)
```


### Load the historical occurrence data and explore patterns
```{r}
hist_occ <- read.csv("data/historical-occurrence-data.csv")

# count of data by year
occ_year <- hist_occ %>%
  group_by(year) %>%
  count()

# The above count suggests that majority of the bird specimens were recorded in 1881, followed by some data from 1876. 

# Let us create a new column that bins year by time period. For the sake of exploratory analysis, we bin data into 50 year time periods (1850-1900, 1900-1950, 1950-2000, post2000)

hist_occ <- hist_occ %>%
  mutate(timePeriod = case_when(year>=1850 & year<1900 ~ "1850-1900",
                                year>=1900 & year<1950 ~ "1900-1950",
                                year>=1950 & year<2000 ~ "1950-2000",
                                year>=2000 ~ "post2000"))

# count by timePeriod
occ_timePeriod <- hist_occ %>%
  group_by(timePeriod) %>%
  count()

# Majority of the data is from the period 1850-1900. 
fig_timePeriod <- ggplot(occ_timePeriod, aes(x=timePeriod, y=n, fill=timePeriod))    + geom_bar(stat="identity", position=position_dodge(), fill="#883107", alpha=0.9) +
  geom_text(aes(label=n, hjust="middle", vjust=-0.5), 
            position = position_dodge(), angle=0, size=5) +
    theme_bw() +
    labs(x="\nTime Period", 
       y="Count\n") +
    theme(axis.title = element_text(family = "Century Gothic",
      size = 14, face = "bold"),
        axis.text = element_text(family="Century Gothic",size = 14),
        legend.position = "none")

ggsave(fig_timePeriod, filename = "figs/fig_specimenCountBytimePeriod.png", width=12, height=7,device = png(), units="in", dpi = 300); dev.off()
```

### Examining count of specimens by locality and time period
```{r}
occ_locality <- hist_occ %>%
  group_by(historical_site_code) %>%
  count()

# Top 5 locations are OTCM, CNRG, LWSL, AVLC and KJPN.

loc_timePeriod <- hist_occ %>%
  group_by(timePeriod, historical_site_code) %>%
  count()

fig_loc_timePeriod <- ggplot(loc_timePeriod, aes(x=historical_site_code, y=n, fill=historical_site_code)) + 
  geom_bar(stat="identity", position=position_dodge(), fill="#883107", alpha=0.9) +
  geom_text(aes(label=n, hjust="middle", vjust=-0.5), 
            position = position_dodge(), angle=0, size=5) +
  facet_wrap(~timePeriod) +
    theme_bw() +
    labs(x="\nTime Period & Locality", 
       y="Count\n") +
    theme(axis.title = element_text(family = "Century Gothic",
      size = 14, face = "bold"),
        axis.text = element_text(family="Century Gothic",size = 14),
      axis.text.x = element_text(angle=90, vjust=0.5, hjust=1),
        legend.position = "none")

ggsave(fig_loc_timePeriod, filename = "figs/fig_specimenCount_locality_timePeriod.png", width=12, height=7,device = png(), units="in", dpi = 300); dev.off()

```

### Examining species specific counts by timePeriod and locality
```{r}
spp_timePeriod <- hist_occ %>%
  group_by(species, timePeriod) %>%
  count() 

# Plotting locations that had a min of 10 species observations per timePeriod

fig_spp_timePeriod <- ggplot(spp_timePeriod, aes(x=species, y=n, fill=species)) + geom_bar(stat="identity", position=position_dodge(), fill="#883107", alpha=0.9) +
  geom_text(aes(label=n, hjust="middle", vjust=-0.5), 
            position = position_dodge(), angle=0, size=5) +
  facet_wrap(~timePeriod) +
    theme_bw() +
    labs(x="\nTime Period & Species", 
       y="Count\n") +
    theme(axis.title = element_text(family = "Century Gothic",
      size = 14, face = "bold"),
        axis.text = element_text(family="Century Gothic",size = 14),
      axis.text.x = element_text(angle=90, vjust=0.5, hjust=1),
        legend.position = "none")

ggsave(fig_spp_timePeriod, filename = "figs/fig_species_timePeriod.png", width=50, height=20, device = png(), units="in", dpi = 300, limitsize = F); dev.off()

# Examining both species by time period and locality
spp_loc_timePeriod <- hist_occ %>%
  group_by(species, timePeriod, historical_site_code) %>%
  count()%>%
  filter(n>10)

fig_spp_loc_timePeriod <- ggplot(spp_loc_timePeriod, aes(x=species, y=n, fill=species)) + geom_bar(stat="identity", position=position_dodge(), fill="#883107", alpha=0.9) +
  geom_text(aes(label=n, hjust="middle", vjust=-0.5), 
            position = position_dodge(), angle=0, size=5) +
  facet_wrap(~timePeriod+historical_site_code) +
    theme_bw() +
    labs(x="\nTime Period, Species & Locality", 
       y="Count\n") +
    theme(axis.title = element_text(family = "Century Gothic",
      size = 14, face = "bold"),
        axis.text = element_text(family="Century Gothic",size = 14),
      axis.text.x = element_text(angle=90, vjust=0.5, hjust=1),
        legend.position = "none")

ggsave(fig_spp_loc_timePeriod, filename = "figs/fig_species_loc_timePeriod_min10.png", width=15, height=10,device = png(), units="in", dpi = 300); dev.off()
```
