==> 02_land-cover-analysis.Rmd <==
---
editor_options: 
  chunk_output_type: console
---

# Land cover data analysis

In this script, we will analyze land cover changes for multiple land cover classes between 1870 and 2017.  

### Load necessary libraries
```{r}
library(sf)
library(raster)
library(terra)
library(stars)
library(dplyr)
library(tidyverse)
library(mapview)
library(landscapemetrics)
library(scico)
```

### Load the Nilgiris contour 

This is the contour to which all other shapefiles will be clipped to and is essentially a 1400 m contour. 
```{r}
# load the 1400m contour
contourNil <- st_read("data/spatial/contour-1400m-nilgiris.shp")
```

### Loading digitized shapefiles of the 1870 Ouchterlony map

The 1870 published maps were based on a survey carried out in 1848 by J. Ouchterlony.  
```{r}
# list all shapefiles in the directory
nilgiris1870 <- list.files("data/spatial/1870-nilgiris/", full.names = T,
                    recursive = T, pattern=".shp$")

# create vector files
cultivated_nil1870 <- st_read(nilgiris1870[1])
forest_nil1870 <- st_read(nilgiris1870[2])
grassland_nil1870 <- st_read(nilgiris1870[3])
plantations_nil1870 <- st_read(nilgiris1870[4])
roads_nil1870 <- st_read(nilgiris1870[5])
settlements_nil1870 <- st_read(nilgiris1870[6])
waterBodies_nil1870 <- st_read(nilgiris1870[7])

# explore and fix any issues with the above vector files

# fix class names and remove empty geometries
plantations_nil1870$class <- "plantations"
settlements_nil1870 <- settlements_nil1870[-371, ]

# Note: the roads shapefile is a linestring while the other geometry types are polygon

# transform to UTM 43N
cultivated_nil1870 <- st_transform(cultivated_nil1870, 32643)
forest_nil1870 <- st_transform(forest_nil1870, 32643)
grassland_nil1870 <- st_transform(grassland_nil1870, 32643)
plantations_nil1870 <- st_transform(plantations_nil1870, 32643)
settlements_nil1870 <- st_transform(settlements_nil1870, 32643)
waterBodies_nil1870 <- st_transform(waterBodies_nil1870, 32643)

# some more processing to ensure class names are standardized across time periods
forest_nil1870$class <- "Shola Forest"
# grassland_nil1870 <- st_cast(grassland_nil1870, "POLYGON")
grassland_nil1870$class <- "Shola Grassland"
waterBodies_nil1870$class <- "waterbodies"
```

### Loading the Kundah 1870 digitized shapefiles

These shapefiles have been digitized from a map of the Koondah mountains (based on a survey by J. Ouchterlony). 
```{r}
# load the Kundah 1870 shapefiles
kundah <- list.files("data/spatial/1870-kundah/",
  full.names = T,
  recursive = T, pattern = ".shp$"
)

# create vector files
forest_kun1870 <- st_read(kundah[1])
grassland_kun1870 <- st_read(kundah[2])
settlements_kun1870 <- st_read(kundah[3])
waterbodies_kun1870 <- st_read(kundah[4])
# whitebars_kun1870 <- st_read(kundah[5])

# explore and fix any issues with the above vector files
names(forest_kun1870) <- c("id", "class", "gridcode", "geometry")
settlements_kun1870$class <- "settlements"

# transform to UTM 43N
forest_kun1870 <- st_transform(forest_kun1870, 32643)
grassland_kun1870 <- st_transform(grassland_kun1870, 32643)
settlements_kun1870 <- st_transform(settlements_kun1870, 32643)
waterbodies_kun1870 <- st_transform(waterbodies_kun1870, 32643)

# some more processing to ensure class names are standardized across time periods
forest_kun1870$class <- "Shola Forest"
#grassland_kun1870 <- st_cast(grassland_kun1870, "POLYGON")
grassland_kun1870$class <- "Shola Grassland"
```

### Combining/unionizing geometries across the Nilgiris 1870 and the Kundah 1870 map

In this script, we unionize geometries across the above two digitized maps that were loaded to ensure that overlap in certain areas is counted as a single entity. This is being done to include the Kundah map in the final analysis.  
```{r}
# the classes that are common across the Nilgiris 1870 and the Kundah 1870 are:
# Shola forests
# Shola grasslands
# Settlements 
# Water bodies

# Unionizing the above features
# forests
forest1870 <- rbind(forest_nil1870, forest_kun1870)
forest1870 <- st_buffer(forest1870, dist=0)
forest1870 <- st_union(forest1870)

# grasslands
grasslands1870 <- rbind(grassland_nil1870, grassland_kun1870)
grasslands1870 <- st_buffer(grasslands1870, dist=0)
grasslands1870 <- st_union(grasslands1870)

# settlements
settlements1870 <- rbind(settlements_nil1870, settlements_kun1870)
settlements1870 <- st_buffer(settlements1870, dist=0)
settlements1870 <- st_union(settlements1870)

# water bodies
waterBodies1870 <- rbind(waterBodies_nil1870, waterbodies_kun1870)
waterBodies1870 <- st_buffer(waterBodies1870, dist=0)
waterBodies1870 <- st_union(waterBodies1870)

# converting geometries back to simple feature collections/dataframes
# gridcode/id does not matter for the purpose of our analysis
# Forests
forestnil1870 <- st_sf(geometry = forest1870)
forestnil1870$class <- "Shola Forest"
forestnil1870$id <- 1:length(forest1870)
forestnil1870$gridcode <- 1:length(forest1870)

# grasslands
grasslandsnil1870 <- st_sf(geometry = grasslands1870)
grasslandsnil1870$class <- "Shola Grassland"
grasslandsnil1870$id <- 1:length(grasslands1870)
grasslandsnil1870$gridcode <- 1:length(grasslands1870)

# settlements
settlementsnil1870 <- st_sf(geometry = settlements1870)
settlementsnil1870$class <- "settlements"
settlementsnil1870$id <- 1:length(settlements1870)
settlementsnil1870$gridcode <- 1:length(settlements1870)

# water bodies
waterBodiesnil1870 <- st_sf(geometry = waterBodies1870)
waterBodiesnil1870$class <- "waterbodies"
waterBodiesnil1870$id <- 1:length(waterBodies1870)
waterBodiesnil1870$gridcode <- 1:length(waterBodies1870)

# creating a single simple feature collection
nil1870 <- rbind(
  cultivated_nil1870, forestnil1870, grasslandsnil1870,
  plantations_nil1870, settlementsnil1870, waterBodiesnil1870
)

# crop the 1870 shapefiles to within the 1400m contour only
nil1870 <- st_buffer(nil1870, dist = 0)
nil1870 <- st_intersection(nil1870, contourNil)
nil1870 <- nil1870[, -c(4, 5, 6)]

# the above steps have resulted in a single dataframe for the 1400m contour only

# Combine the nilgiris 1870 (Ouchterlony map) and the kundah 1870 for clipping
# with files from other time periods (see below)
all1870 <- st_buffer(st_union(nil1870), dist = 0)
```


### Load the shapefiles from 1973, 1995 and 2017 for the Nilgiris
```{r}
# load the 1973, 1995 and 2017 shapefiles
nil1973 <- st_read("data/spatial/1973-nilgiris/1973D.shp")
nil1995 <- st_read("data/spatial/1995-nilgiris/1995D.shp")
nil2017 <- st_read("data/spatial/2017-nilgiris/2017D.shp")

# Each of the above shapefiles have slightly different characteristics
# class name is present only in nil2017 and not in the others
# nil 1973 and nil 1995 are of geometry type MULITPOLYGON while nil 2017 is of geometry type POLYGON

# Let's first add class name to nil1973 shapefile
nil1973 <- nil1973 %>%
  mutate(class = case_when(
    gridcode == 1 ~ "Shola Grassland",
    gridcode == 2 ~ "Shola Forest",
    gridcode == 3 ~ "Timber Plantations",
    gridcode == 4 ~ "Tea Plantations",
    gridcode == 5 ~ "Ochlandra",
    gridcode == 6 ~ "Settlements",
    gridcode == 7 ~ "Agricultural Land",
    gridcode == 8 ~ "Water bodies"
  ))

# Let's add class name to nil1995 shapefile
nil1995 <- nil1995 %>%
  mutate(class = case_when(
    gridcode == 1 ~ "Shola Grassland",
    gridcode == 2 ~ "Shola Forest",
    gridcode == 3 ~ "Timber Plantations",
    gridcode == 4 ~ "Tea Plantations",
    gridcode == 5 ~ "Ochlandra",
    gridcode == 6 ~ "Settlements",
    gridcode == 7 ~ "Agricultural Land",
    gridcode == 8 ~ "Water bodies"
  ))

# crop the 1973, 1995 and 2017 shapefiles to within the 1400m contour only
nil1973 <- st_intersection(nil1973, all1870)
nil1995 <- st_intersection(nil1995, all1870)
nil2017 <- st_intersection(nil2017, all1870)
```

### Rasterization of vector files

The decision to use a vector or raster for further analysis depends on the question being asked. Please refer to script in code/01_vectorVsRaster.R  
```{r}
# nil1870 raster
vectnil1870 <- terra::vect(nil1870)
emptyRast <- terra::rast(res = 12, xmin = 655224.34214709, xmax = 718286.685853398, ymin = 1229957.78965221, ymax = 1275965.51753469, crs = "+proj=utm +zone=43 +datum=WGS84 +units=m +no_defs")
rastnil1870 <- terra::rasterize(vectnil1870, emptyRast, "class")

# Let's load the 1973, 1995 and 2017 series and rasterize them
vectnil1973 <- vect(nil1973)
emptyRast1973 <- terra::rast(res = 30, xmin = 655224.34214709, xmax = 718286.685853398, ymin = 1229957.78965221, ymax = 1275965.51753469, crs = "+proj=utm +zone=43 +datum=WGS84 +units=m +no_defs")
rastnil1973 <- terra::rasterize(vectnil1973, emptyRast1973, "class")
sz1973 <- cellSize(rastnil1973, unit = "m")
areaRast1973 <- zonal(sz1973, rastnil1973, sum)
areaRast1973SqKm <- (areaRast1973$area / 1000000)
areaRast1973 <- cbind(areaRast1973, areaRast1973SqKm)
names(areaRast1973) <- c("class", "areaInMeters", "sumArea")

# 1995
vectnil1995 <- vect(nil1995)
emptyRast1995 <- terra::rast(res = 30, xmin = 655224.34214709, xmax = 718286.685853398, ymin = 1229957.78965221, ymax = 1275965.51753469, crs = "+proj=utm +zone=43 +datum=WGS84 +units=m +no_defs")
rastnil1995 <- terra::rasterize(vectnil1995, emptyRast1995, "class")
sz1995 <- cellSize(rastnil1995, unit = "m")
areaRast1995 <- zonal(sz1995, rastnil1995, sum)
areaRast1995SqKm <- (areaRast1995$area / 1000000)
areaRast1995 <- cbind(areaRast1995, areaRast1995SqKm)
names(areaRast1995) <- c("class", "areaInMeters", "sumArea")

# 2017
vectnil2017 <- vect(nil2017)
emptyRast2017 <- terra::rast(res = 30, xmin = 655224.34214709, xmax = 718286.685853398, ymin = 1229957.78965221, ymax = 1275965.51753469, crs = "+proj=utm +zone=43 +datum=WGS84 +units=m +no_defs")
rastnil2017 <- terra::rasterize(vectnil2017, emptyRast2017, "class")
sz2017 <- cellSize(rastnil2017, unit = "m")
areaRast2017 <- zonal(sz2017, rastnil2017, sum)
areaRast2017SqKm <- (areaRast2017$area / 1000000)
areaRast2017 <- cbind(areaRast2017, areaRast2017SqKm)
names(areaRast2017) <- c("class", "areaInMeters", "sumArea")

# resampling rasters to ensure they are all of the same spatial resolution and extent
# We have to do this for the nil1870 and kun1870 rasters

resamNil1870 <- terra::resample(rastnil1870, rastnil1973, method = "near")
sz1870 <- cellSize(resamNil1870, unit = "m")
areaNil1870 <- zonal(sz1870, resamNil1870, sum)
areaNil1870SqKm <- (areaNil1870$area / 1000000)
areaNil1870 <- cbind(areaNil1870, areaNil1870SqKm)
```

### Cleaning up the 1870 map
```{r}
# 1870 map
areaNil1870$class <- c(
  "Shola Forest", "Shola Grassland",
  "Agricultural Land", "Plantations (Tea & Timber)",
  "Settlements", "Water bodies"
)
```

### Calculating areas of different land cover classes across time periods using rasters

Making plots of the overall area of land cover types
```{r}
# Joining all dataframes to create a single one for plotting and visualization
areaCalc <- purrr::reduce(list(
  areaNil1870, areaRast1973, areaRast1995,
  areaRast2017
), dplyr::full_join, by = "class")

names(areaCalc) <- c("class", "areaMt1870", "1870", "areaMt1973", "1973", "areaMt1995", 
                     "1995", "areaMt2017", "2017")

areaCalc <- areaCalc %>%
  select("class", "1870", "1973", "1995", "2017") %>%
  pivot_longer(!class, names_to = "Year", values_to = "Area")

write.csv(areaCalc, "results/totalArea-by-landCover-timePeriod.csv", row.names = F)

# make plot
fig_area <- ggplot(areaCalc, aes(x = class, y = Area, fill = class)) +
  geom_bar(stat = "identity", position = position_dodge(), fill = "#883107", alpha = 0.9) +
  geom_text(aes(label = round(Area), hjust = "middle", vjust = -0.5),
    position = position_dodge(), angle = 0, size = 5
  ) +
  scale_fill_scico_d(palette = "batlow") +
  facet_wrap(~Year) +
  theme_bw() +
  labs(
    x = "\nLand cover type",
    y = "Area in sq.km. \n"
  ) +
  theme(
    axis.title = element_text(
      family = "Century Gothic",
      size = 14, face = "bold"
    ),
    axis.text = element_text(family = "Century Gothic", size = 14),
    axis.text.x = element_text(angle = 90, vjust = 0.5, hjust = 1),
    legend.position = "none"
  )

ggsave(fig_area,
  filename = "figs/fig_totalArea_landCover.png", width = 15, height = 13,
  device = png(), units = "in", dpi = 300
)
```

### Create a figure of proportion of landcover across areas over time (using rasters)
```{r}
# get percentArea occupied by each land cover class
percentArea <- areaCalc %>%
  group_by(Year) %>%
  mutate(
    totalArea = sum(Area, na.rm = T),
    percentArea = (Area / totalArea) * 100
  )

# plot figure
fig_percent_area <- ggplot(percentArea, aes(x = class, y = percentArea, fill = class)) +
  geom_bar(stat = "identity", position = position_dodge(), alpha = 0.9) +
  geom_text(aes(label = round(percentArea, digits = 1), hjust = "middle", vjust = -0.5), position = position_dodge(), angle = 0, size = 5) +
  scale_fill_scico_d(palette = "batlow") +
  facet_wrap(~Year) +
  theme_bw() +
  labs(
    x = "\nLand cover type",
    y = "Percent Area \n"
  ) +
  theme(
    axis.title = element_text(
      family = "Century Gothic",
      size = 14, face = "bold"
    ),
    axis.text = element_text(family = "Century Gothic", size = 14),
    axis.text.x = element_text(angle = 90, vjust = 0.5, hjust = 1),
    legend.position = "none"
  )

ggsave(fig_percent_area, filename = "figs/fig_percentArea_landCover.png", width = 14, height = 13, device = png(), units = "in", dpi = 300)
```

## Re-running the area calculations with vector files

This calculation is being run to assess if the area-wise calculations of changes when vector files are being used. 
```{r}
# 1870 
areaVect1870 <- nil1870 %>%
  mutate(areaSqKm = as.numeric(st_area(nil1870)/1000000)) # in sq. km.

sumArea1870 <- areaVect1870 %>%
  group_by(class) %>%
  mutate(sumArea = sum(areaSqKm)) %>%
  distinct(class, sumArea)

# 1973
areaVect1973 <- nil1973 %>%
  mutate(areaSqKm = as.numeric(st_area(nil1973)/1000000)) # in sq. km.

sumArea1973 <- areaVect1973 %>%
  group_by(class) %>%
  mutate(sumArea = sum(areaSqKm)) %>%
  distinct(class, sumArea)

# 1995
areaVect1995 <- nil1995 %>%
  mutate(areaSqKm = as.numeric(st_area(nil1995)/1000000)) # in sq. km.

sumArea1995 <- areaVect1995 %>%
  group_by(class) %>%
  mutate(sumArea = sum(areaSqKm)) %>%
  distinct(class, sumArea)

# 1973
areaVect2017 <- nil2017 %>%
  mutate(areaSqKm = as.numeric(st_area(nil2017)/1000000)) # in sq. km.

sumArea2017 <- areaVect2017 %>%
  group_by(class) %>%
  mutate(sumArea = sum(areaSqKm)) %>%
  distinct(class, sumArea)

# fix 1870 class names
sumArea1870$class <- c("Agricultural Land",
  "Shola Forest", "Shola Grassland",
   "Plantations (Tea & Timber)",
  "Settlements", "Water bodies"
)
```

### Calculating areas of different land cover classes across time periods (vector files)

Making plots of the overall area of land cover types
```{r}
# Joining all dataframes to create a single one for plotting and visualization
areaCalc <- purrr::reduce(list(
  sumArea1870, sumArea1973, sumArea1995,
  sumArea2017
), dplyr::full_join, by = "class")

names(areaCalc) <- c("class", "1870", "1973","1995", "2017")

areaCalc <- areaCalc %>%
  select("class", "1870", "1973", "1995", "2017") %>%
  pivot_longer(!class, names_to = "Year", values_to = "Area")

write.csv(areaCalc, "results/totalArea-by-landCover-timePeriod-vector.csv", row.names = F)

# make plot
fig_area <- ggplot(areaCalc, aes(x = class, y = Area, fill = class)) +
  geom_bar(stat = "identity", position = position_dodge(), fill = "#883107", alpha = 0.9) +
  geom_text(aes(label = round(Area), hjust = "middle", vjust = -0.5),
    position = position_dodge(), angle = 0, size = 5
  ) +
  scale_fill_scico_d(palette = "batlow") +
  facet_wrap(~Year) +
  theme_bw() +
  labs(
    x = "\nLand cover type",
    y = "Area in sq.km. \n"
  ) +
  theme(
    axis.title = element_text(
      family = "Century Gothic",
      size = 14, face = "bold"
    ),
    axis.text = element_text(family = "Century Gothic", size = 14),
    axis.text.x = element_text(angle = 90, vjust = 0.5, hjust = 1),
    legend.position = "none"
  )

ggsave(fig_area,
  filename = "figs/fig_totalArea_landCover-vector.png", width = 15, height = 13,
  device = png(), units = "in", dpi = 300
)
```

### Create a figure of proportion of landcover across areas over time
```{r}
# get percentArea occupied by each land cover class
percentArea <- areaCalc %>%
  group_by(Year) %>%
  mutate(
    totalArea = sum(Area, na.rm = T),
    percentArea = (Area / totalArea) * 100
  )

# plot figure
fig_percent_area <- ggplot(percentArea, aes(x = class, y = percentArea, fill = class)) +
  geom_bar(stat = "identity", position = position_dodge(), alpha = 0.9) +
  geom_text(aes(label = round(percentArea, digits = 1), hjust = "middle", vjust = -0.5), position = position_dodge(), angle = 0, size = 5) +
  scale_fill_scico_d(palette = "batlow") +
  facet_wrap(~Year) +
  theme_bw() +
  labs(
    x = "\nLand cover type",
    y = "Percent Area \n"
  ) +
  theme(
    axis.title = element_text(
      family = "Century Gothic",
      size = 14, face = "bold"
    ),
    axis.text = element_text(family = "Century Gothic", size = 14),
    axis.text.x = element_text(angle = 90, vjust = 0.5, hjust = 1),
    legend.position = "none"
  )

ggsave(fig_percent_area, filename = "figs/fig_percentArea_landCover-vector.png", width = 14, height = 13, device = png(), units = "in", dpi = 300)
```


1. total area by land cover + time Period - DONE
2. Proportion of the landscape occupied by each landcover over time - DONE
2. pathc-level area analysis by land cover + time Period
3. heterogeneity of the land cover data using vegan::diversity() (Auffret et al. 2018)
