==> 07_occurrence-exploratory-analyses.Rmd <==
---
editor_options: 
  chunk_output_type: console
---

# Exploratory analysis of occurrence data

This script carries out exploratory analysis of historical and modern occurrence data from the Nilgiri hills. In this script, we plot the total count of specimens by time period and the number of specimens per species by time period.   

### Load necessary libraries
```{r}
library(dplyr)
library(stringr)
library(tidyverse)
library(scico)
library(RColorBrewer)
library(extrafont)
```

### Load the historical occurrence data and explore patterns
```{r}
hist_occ <- read.csv("data/historical-occurrence-data.csv")

# count of data by year
occ_year <- hist_occ %>%
  group_by(year) %>%
  count()

# The above count suggests that majority of the bird specimens were recorded in 1881, followed by some data from 1876.

# For further downstream analysis, we ignore data from 1901 onwards
hist_occ <- hist_occ %>%
  filter(year <= 1901)
```

### Examining count of specimens by locality
```{r}
# ten unique historical site localities across which species were recorded
occ_locality <- hist_occ %>%
  group_by(historical_site_code) %>%
  count()

# Examining species by locality
spp_loc <- hist_occ %>%
  group_by(scientific_name, historical_site_code) %>%
  count()

# how many unique sites
prop_count <- spp_loc %>%
  group_by(scientific_name) %>%
  summarise(uniqueSites = length(unique(historical_site_code)))

## count unique species list
spp_count <- hist_occ %>%
  group_by(scientific_name) %>%
  count()

unique(spp_count$scientific_name) # a total of 148 species were collected/recorded between 1850 and 1901 across 12 unique historical localities identified
```

## Visualizations of historical data
```{r}
fig_loc_timePeriod <- ggplot(occ_locality, aes(x = reorder(historical_site_code,-n), y = n, fill = historical_site_code)) +
  geom_bar(stat = "identity", position = position_dodge(), fill = "#883107", alpha = 0.9) +
  geom_text(aes(label = n, hjust = "middle", vjust = -0.5),
    position = position_dodge(), angle = 0, size = 5
  ) +
  theme_bw() +
  labs(
    x = "\nLocality",
    y = "Count\n"
  ) +
  theme(
    axis.title = element_text(
      family = "Century Gothic",
      size = 14, face = "bold"
    ),
    axis.text = element_text(family = "Century Gothic", size = 14),
    axis.text.x = element_text(angle = 90, vjust = 0.5, hjust = 1),
    legend.position = "none"
  )

ggsave(fig_loc_timePeriod, filename = "figs/fig_histCount_locality.png", width = 12, height = 7, device = png(), units = "in", dpi = 300)
dev.off()
```


## Examining species specific counts by locality
```{r}
fig_spp <- ggplot(spp_count, aes(x = reorder(scientific_name, -n), y = n, fill = scientific_name)) +
  geom_bar(stat = "identity", position = position_dodge(), fill = "#883107", alpha = 0.9) +
  geom_text(aes(label = n, hjust = "middle", vjust = -0.5),
    position = position_dodge(), angle = 0, size = 5
  ) +
  theme_bw() +
  labs(
    x = "\nSpecies",
    y = "Count\n"
  ) +
  theme(
    axis.title = element_text(
      family = "Century Gothic",
      size = 14, face = "bold"
    ),
    axis.text = element_text(family = "Century Gothic", size = 14),
    axis.text.x = element_text(angle = 90, vjust = 0.5, hjust = 1),
    legend.position = "none"
  )

ggsave(fig_spp, filename = "figs/fig_histSpp.png", width = 50, height = 20, device = png(), units = "in", dpi = 300, limitsize = F)
dev.off()
```

## Load modern occurrence data and explore patterns
```{r}
mod_occ <- read.csv("data/modern-occurrence-data.csv") %>%
  filter(!is.na(historical_site_code)) %>%
  filter(!is.na(species_code))

# total observations of species in the modern surveys
modOcc_spp <- mod_occ %>%
  group_by(species_code) %>%
  count() # 2301 observations of 102 bird species

# plot figure of species by locality
mod_occ_loc <- mod_occ %>%
  group_by(scientific_name, historical_site_code) %>%
  count()

# figure of species by locality
fig_modSpp_loc <- ggplot(mod_occ_loc, aes(x = reorder(historical_site_code,-n), y = n, fill = historical_site_code)) +
  geom_bar(stat = "identity", position = position_dodge(), fill = "#883107", alpha = 0.9) +
  geom_text(aes(label = n, hjust = "middle", vjust = -0.5),
    position = position_dodge(), angle = 0, size = 5
  ) +
  facet_wrap(~scientific_name) +
  theme_bw() +
  labs(
    x = "\nSpecies and Locality",
    y = "Count\n"
  ) +
  theme(
    axis.title = element_text(
      family = "Century Gothic",
      size = 14, face = "bold"
    ),
    axis.text = element_text(family = "Century Gothic", size = 12),
    axis.text.x = element_text(angle = 90, vjust = 0.5, hjust = 1),
    legend.position = "none"
  )

ggsave(fig_modSpp_loc, filename = "figs/fig_modSpp_loc.png", width = 40, height = 10, device = png(), units = "in", dpi = 300)
dev.off()

## Some interesting patterns of certain species being more abundant/detected in certain sites over others.
```
