==> 08_historical-modern-occurrence-comparisons.Rmd <==
---
editor_options: 
  chunk_output_type: console
---

## Historical-modern occurrence comparisons

Here, we will compare historical and modern occurrence data and test for differences in presences/absences of species across time periods and sites

### Load necessary libraries

```{r}
library(dplyr)
library(stringr)
library(tidyverse)
library(scico)
library(RColorBrewer)
library(extrafont)
library(sf)
library(raster)
```

### Load list of resurvey locations

We will load the list of sites across which a modern resurvey was carried out, along with

```{r}
# load list of resurvey locations and add elevation as a variable
# remove the modern resurvey locations only (ie. ASFQ)
sites <- read.csv("data/list-of-resurvey-locations.csv")
sites <- sites[-c(1:50),]

# convert to sf object and transform
sites <- st_as_sf(sites, coords = c("longitude", "latitude")) %>%
  `st_crs<-`(4326) %>%
  st_transform(32643)

# add elevation raster
alt <- raster("data/spatial/elevation/alt") # this layer is not added to github as a result of its large size and can be downloaded from SRTM (Farr et al. (2007))

# extract values from that raster (note: transformation of coordinate system)
elev <- extract(alt, sites)
sites <- cbind(sites, elev)
```

### Load historical and modern occurrence data

For the historical occurrence data, we will choose data from the time period 1850-1900 as majority of specimen collections were carried out in this time period. Further below, we will also choose data from 1900-1950 and 1950-2000 to explore comparisons with modern occurrence data.

```{r}
## read in historical occurrence data
hist_occ <- read.csv("data/historical-occurrence-data.csv")

### pre-1900
hist_occ_pre1900 <- hist_occ %>%
  filter(year <= 1900)

## filter by site and get unique species per site (historical)
histOcc_spp_pre1900 <- hist_occ_pre1900 %>%
  group_by(scientific_name, historical_site_code) %>%
  count()

colnames(histOcc_spp_pre1900)[3] <- "histCount"

# add elevation
hist_occ_elev_pre1900 <-  left_join(histOcc_spp_pre1900, sites, by=c("historical_site_code"="historical_site_code")) %>%
  filter(!is.na(scientific_name)) %>%
  replace(is.na(.), 0) %>%
  mutate("1850-1900" = case_when(histCount >= 1 ~ 1,TRUE ~ 0))

# repeat above for two more time-periods
## 1900-1950
hist_occ_1900_1950 <- hist_occ %>%
  filter(year > 1900 & year <= 1950)

## filter by site and get unique species per site (historical)
histOcc_spp_1900_1950 <- hist_occ_1900_1950 %>%
  group_by(scientific_name, historical_site_code) %>%
  count()

colnames(histOcc_spp_1900_1950)[3] <- "histCount"

# add elevation
hist_occ_elev_1900_1950 <-  left_join(histOcc_spp_1900_1950, sites, by=c("historical_site_code"="historical_site_code")) %>%
  filter(!is.na(scientific_name)) %>%
  replace(is.na(.), 0) %>%
  mutate("1900-1950" = case_when(histCount >= 1 ~ 1,TRUE ~ 0))

### 1950 - 2000
hist_occ_1950_2000 <- hist_occ %>%
  filter(year > 1950 & year <= 2000)

## filter by site and get unique species per site (historical)
histOcc_spp_1950_2000 <- hist_occ_1950_2000 %>%
  group_by(scientific_name, historical_site_code) %>%
  count()

colnames(histOcc_spp_1950_2000)[3] <- "histCount"

# add elevation
hist_occ_elev_1950_2000 <-  left_join(histOcc_spp_1950_2000, sites, by=c("historical_site_code"="historical_site_code")) %>%
  filter(!is.na(scientific_name)) %>%
  replace(is.na(.), 0) %>%
  mutate("1950-2000" = case_when(histCount >= 1 ~ 1,TRUE ~ 0))

## read in modern occurrence data
mod_occ <- read.csv("data/modern-occurrence-data.csv")

## filter by site and get unique species per site (modern)
modOcc_spp <- mod_occ %>%
  filter(!is.na(historical_site_code)) %>%
  group_by(scientific_name, modern_site_code) %>%
  summarise(modCount = sum(number))

mod_occ_elev <-  left_join(modOcc_spp, sites, by=c("modern_site_code"="modern_site_code")) %>%
  filter(!is.na(scientific_name)) %>%
  replace(is.na(.), 0) %>%
  mutate("2021" = case_when(modCount >= 1 ~ 1,TRUE ~ 0))
```

### Comparing historical-modern counts by elevation

```{r}
## joining the historical datasets
hist_mod <- full_join(hist_occ_elev_pre1900, hist_occ_elev_1900_1950, by = c("scientific_name","modern_site_code","histCount","historical_site_code","elev","site_name","geometry")) %>%
  full_join(., hist_occ_elev_1950_2000, by = c("scientific_name","modern_site_code","histCount","historical_site_code","elev","site_name","geometry")) %>%
   full_join(., mod_occ_elev, by = c("scientific_name","modern_site_code","historical_site_code","elev","site_name","geometry")) %>%
  replace(is.na(.), 0) %>%
  arrange(scientific_name)

# run the loop to create a list of plots
plots <- list()

for(i in 1:length(unique(hist_mod$scientific_name))) {
  
  # extract data species by species
  a <- unique(hist_mod$scientific_name)[i]
  data <- hist_mod[hist_mod$scientific_name==a,] %>%
      pivot_longer(cols=c("1850-1900", "1900-1950", "1950-2000", "2021"), 
                   names_to="histMod", values_to="PresAbs") %>%
    arrange(histMod)
  
  # plot the data
  plots[[i]] <- ggplot(data, aes(x = histMod, y = elev, fill = factor(PresAbs))) +
    geom_point(aes(shape=factor(PresAbs), size = 3)) +
    scale_shape_manual(values = c(1, 16)) +
  theme_bw() +
  ggtitle(a) +
  labs(
    x = "\nHistoric vs. Modern Survey",
    y = "Elevation (in meters)\n"
  ) +
  theme(
    plot.title = element_text(family = "Century Gothic", face="bold.italic"),
    axis.title = element_text(
      family = "Century Gothic",
      size = 14, face = "bold"
    ),
    axis.text = element_text(family = "Century Gothic", size = 13),
    axis.text.x = element_text(angle = 0, vjust = 0.5, hjust = 0.5), 
    legend.title = element_text(family = "Century Gothic"),
    legend.text = element_text(family = "Century Gothic")) + 
    guides(size= "none",
           fill = "none",
           shape = guide_legend(title = "Presence-Absence"))
}

# plot and save as a single pdf
cairo_pdf(
  filename = "figs/elevation-historic-modern-occurrence.pdf",
  onefile = TRUE
)
plots
dev.off()
```
