==> 09_species-relative-abundances-over-time.Rmd <==
---
editor_options: 
  chunk_output_type: console
---

### Species relative abundances over time

We replicate analyses following Gotelli et al. 2021 (who provides a methodological approach to calculate relative species abundances using historical museum records) and Gómez et al. 2021 who assumed two scenarios while calculating extinctions and colonizations - "The first is based on the assumption that all surveys accurately describe the avifauna present and that any records of species not detected in prior surveys of San Antonio represent real colonizations. The second scenario is based on the assumption that all the novel species (i.e., those detected for the first time after the 1910s) were false absences."

The second assumption is that the level of the analysis is at the 'historic-site' and not the 'modern-site' (which essentially corresponds to multiple unique modern survey sites for every 'historic-site'). While calculating relative abundances and for estimation of turnover and to ensure that the analysis is consistent, we sum abundances across the modern sites corresponding to each historic site.  

### Load necessary libraries
```{r}
library(dplyr)
library(stringr)
library(tidyverse)
library(scico)
library(RColorBrewer)
library(extrafont)
library(sf)
library(raster)
library(data.table)

# source custom functions
source("code/02_relative-species-abundance-functions.R")
```

### Load list of resurvey locations

We will load the list of sites across which a modern resurvey was carried out, along with elevation rasters.  
```{r}
# load list of resurvey locations and add elevation as a variable
# remove the modern resurvey locations only (ie. ASFQ)
sites <- read.csv("data/list-of-resurvey-locations.csv")
sites <- sites[-c(1:50),]

# convert to sf object and transform
sites <- st_as_sf(sites, coords = c("longitude", "latitude")) %>%
  `st_crs<-`(4326) %>%
  st_transform(32643)

# add elevation raster
alt <- raster("data/spatial/elevation/alt") # this layer is not added to github as a result of its large size and can be downloaded from SRTM (Farr et al. (2007))

# extract values from that raster (note: transformation of coordinate system)
elev <- extract(alt, sites)
sites <- cbind(sites, elev)
```


For the historical occurrence data, we will choose data from three time periods: 1850-1900, 1900-1950 and 1950-2000. 
```{r}
## read in historical occurrence data
hist_occ <- read.csv("data/historical-occurrence-data.csv")

### pre-1900
hist_occ_pre1900 <- hist_occ %>%
  filter(year <= 1900)

### fix the count column
hist_occ_pre1900 <-  hist_occ_pre1900 %>%
  mutate(histCount = case_when(individualCount == "X" ~ 1,
                               individualCount == "" ~ 1,
                               TRUE ~ as.numeric(individualCount)))

# calculate total abundance of historical data at each site
hist_occ_count_pre1900 <- hist_occ_pre1900 %>%
  group_by(scientific_name, historical_site_code) %>%
  summarise("1850-1900" = sum(histCount)) 

# left join the sites dataframe
hist_occ_sites_pre1900 <- left_join(hist_occ_count_pre1900, sites, by=c("historical_site_code"="historical_site_code")) %>%
  filter(!is.na(scientific_name)) %>%
  replace(is.na(.), 0) 

## 1900-1950
hist_occ_1900_1950 <- hist_occ %>%
  filter(year > 1900 & year <= 1950)

### fix the count column
hist_occ_1900_1950 <-  hist_occ_1900_1950 %>%
  mutate(histCount = case_when(individualCount == "X" ~ 1,
                               individualCount == "" ~ 1,
                               TRUE ~ as.numeric(individualCount)))

# calculate total abundance of historical data at each site
hist_occ_count_1900_1950 <- hist_occ_1900_1950 %>%
  group_by(scientific_name, historical_site_code) %>%
  summarise("1900-1950" = sum(histCount)) 

# left join the sites dataframe
hist_occ_sites_1900_1950 <- left_join(hist_occ_count_1900_1950, sites, by=c("historical_site_code"="historical_site_code")) %>%
  filter(!is.na(scientific_name)) %>%
  replace(is.na(.), 0) 

### 1950 - 2000
hist_occ_1950_2000 <- hist_occ %>%
  filter(year > 1950 & year <= 2000)

### fix the count column
hist_occ_1950_2000 <-  hist_occ_1950_2000 %>%
  mutate(histCount = case_when(individualCount == "X" ~ 1,
                               individualCount == "" ~ 1,
                               TRUE ~ as.numeric(individualCount)))

# calculate total abundance of historical data at each site
hist_occ_count_1950_2000 <- hist_occ_1950_2000 %>%
  group_by(scientific_name, historical_site_code) %>%
  summarise("1950-2000" = sum(histCount)) 

# left join the sites dataframe
hist_occ_sites_1950_2000 <- left_join(hist_occ_count_1950_2000, sites, by=c("historical_site_code"="historical_site_code")) %>%
  filter(!is.na(scientific_name)) %>%
  replace(is.na(.), 0) 

## read in modern occurrence data
mod_occ <- read.csv("data/modern-occurrence-data.csv")

## group by scientific name and modern site code
modOcc_spp <- mod_occ %>%
  filter(!is.na(historical_site_code)) %>%
  filter(!is.na(scientific_name)) %>%
  group_by(scientific_name, modern_site_code) %>%
  summarise("2021" = sum(number))

## left join the sites dataframe
mod_occ_sites <-  left_join(modOcc_spp, sites, by=c("modern_site_code"="modern_site_code")) %>%
  filter(!is.na(scientific_name)) %>%
  replace(is.na(.), 0)
```

### Calculate dirichlet distributions of relative abundance

Here, we will functions from Gotelli et al. 2021 to calculate relative species abundances between historic and modern periods. For the first level of analysis, we ignore information at the site-level and calculate overall relative abundance per time-period at the landscape level. 
```{r}
# note: the level of the analysis is at the 'historic-site' and not the 'modern-site' (which essentially corresponds to three unique sites for every 'historic-site')
# to ensure that the analysis is consistent, we sum abundances across the modern sites corresponding to each historic site
# However, our analysis calculates relative species abundances for the give time period rather than by site + time-period

## calculating relative species abundances for the time period 1850-1900

# we remove modern site_code and get only distinct rows
unique_hist_pre1900 <- hist_occ_sites_pre1900[,c(1,3)] %>%
  group_by(scientific_name) %>%
  summarise("relAbund1850" = sum(`1850-1900`)) 

# apply dirichlet distribution
dat <- tibble::deframe(unique_hist_pre1900) ## choosing 1850-1900 data

relAbun1850 <- dirch_stats(dat) %>%
  dplyr::select(species, bayes_mean) %>%
  rename(., bayes_mean_1850 = bayes_mean)

## calculating relative species abundances for the time period 1900-1950

# we remove modern site_code and get only distinct rows
unique_hist_1900_1950 <- hist_occ_sites_1900_1950[,c(1,3)] %>%
  group_by(scientific_name) %>%
  summarise("relAbund1900" = sum(`1900-1950`)) 

# apply dirichlet distribution
dat <- tibble::deframe(unique_hist_1900_1950) ## choosing 1900-1950 data

relAbun1900 <- dirch_stats(dat) %>%
  dplyr::select(species, bayes_mean) %>%
  rename(., bayes_mean_1900 = bayes_mean)

## calculating relative species abundances for the modern time period

# we remove modern site_code and get only distinct rows
unique_mod <- mod_occ_sites[,c(1,3)] %>%
  group_by(scientific_name) %>%
  summarise("relAbund2021" = sum(`2021`)) 

# apply dirichlet distribution
dat <- tibble::deframe(unique_mod) ## choosing modern data

relAbun2021 <- dirch_stats(dat) %>%
  dplyr::select(species, bayes_mean) %>%
  rename(., bayes_mean_2021 = bayes_mean)

### join the relative abundance datasets together
relAbun <- full_join(relAbun1850, relAbun1900, by = c("species")) %>%
   full_join(., relAbun2021, by = c("species")) %>%
  replace(is.na(.), 0) %>%
  arrange(species) %>%
  rename(., `1850-1900` = "bayes_mean_1850") %>%
  rename(., `1900-1950` = "bayes_mean_1900") %>%
  rename(., `2021` = "bayes_mean_2021")
```

## Visualizing changes in relative abundance over time
```{r}
## This analysis is being done at the landscape-level

## one:one comparisons of abundance between 1850-1900 and 2021 (all species and sites pooled)
fig_1850v2021_rel <- ggplot(data=relAbun,aes(x=`2021`,y=`1850-1900`))  + geom_point() +
   scale_y_log10() +
  geom_abline(slope=1, intercept=0,linetype="dashed") +
  scale_x_log10()+
  labs(y = "1850-1900 Relative abundance", 
       x = "2021 Relative abundance") +
   geom_point(shape = 21, colour = "black", fill = "white", size = 2, stroke = 1)+
   geom_smooth(method="lm", se=TRUE, fullrange=FALSE, level=0.95,linetype="solid") +  theme_bw() +
  theme(axis.text = element_text(family = "Century Gothic", size = 13),
        legend.title = element_text(family = "Century Gothic"),
    legend.text = element_text(family = "Century Gothic"),
    text = element_text(family = "Century Gothic", size = 25))

ggsave(fig_1850v2021_rel, filename = "figs/fig_1850-1900vs2021_relAbun.png", width = 15, height = 10, device = png(), units = "in", dpi = 300)

## boxplots
fig_relAbun_over_time <- relAbun %>%
  dplyr::select(-species) %>%
  pivot_longer(everything()) %>%
  ggplot(aes(x=name, y = value,
             fill = name)) +
  geom_boxplot(alpha=0.7) +
  scale_y_log10() +
  geom_point(size=4,shape=21, alpha=0.5,
             color="black",fill="grey70") +
  geom_jitter(width=0.2,size=4,shape=21, alpha=0.5,
              color="black",fill="grey70") +
  xlab("Time period") +
  ylab("Relative abundance") +
  scale_fill_scico_d(palette = "batlow") +
  theme_bw() +
  theme(axis.text = element_text(family = "Century Gothic", size = 13),
        legend.title = element_text(family = "Century Gothic"),
    legend.text = element_text(family = "Century Gothic"),
    text = element_text(family = "Century Gothic", size = 25)) +
  guides(fill="none")

ggsave(fig_relAbun_over_time, filename = "figs/fig_relAbun_landscapeLevel.png", width = 15, height = 10, device = png(), units = "in", dpi = 300)
```

## Examining changes in relative abundance as a function of species habitat affiliation

#### Let's first curate the species trait data
```{r}
## load state of India's birds data to get habitat affiliation
species_codes <- read.csv("data/species-codes.csv")
habitat_SoIB <- read.csv("data/SoIB-data.csv") # note: we used two files - one file from 2020 and an updated habitat specialization file from 2022 (see below)

# join the two dataframes
habitat_dat <- left_join(species_codes, habitat_SoIB, by=c("scientific_name"="scientific_name"))

# write the above to data folder and recode habitat affiliation
# write.csv(habitat_dat, "data/species-trait-dat.csv", row.names = F)

# note: the classification of habitat type was taken from a working file of trait data for State of India's Birds report v2 (to be released in 2023). We used the following trait classifications:
# Wooded habitat: If it preferred some level of wooded habitat but not particularly specialized on rainforest/forest habitat. An example in this case is the black and orange flycatcher. While it is commonly found in shola forest habitat, this species is also common in degraded wooded areas, including timber plantations. If a species preferred just scrub habitat alone, we lumped it under this category.
# Forest: If a species was a forest specialist. 
# Grassland: While State of India's birds uses a separate classification for grassland & grassland+scrub, here we combine the two and use a single classification for Grassland. Eg. Black drongo - which favors both grassland areas and scrub habitat. There are four species that are classified as using 'open habitat' which had been subsumed under grasslands in our study - this includes - blue-tailed bee-eater, black-winged kite, long-tailed shrike (eBird lists it as preferring open habitats) and western yellow wagtail. 
# Generalist: If a spedcies is truly a generalist species and is not particularly occupying/specialized on any particular wooded/forested/grassland habitat.

## load the species trait data (that includes habitat) and merge with the AVONET dataset
## please note that it is being loaded after curating and adding a column for habitat type based on our understanding of species natural history
hab_dat <- read.csv("data/species-habitat-dat.csv")
avonet <- read.csv("data/AVONET-data.csv")

# join the above two datasets
trait_dat <- left_join(hab_dat, avonet[,c(1,10:20,29:31)], by=c("scientific_name"="Species2"))

## Write the trait data to file
## write.csv(trait_dat,"data/species-trait-dat.csv",row.names = F)
```

## Join datasets of relative abundance and species trait
```{r}
## load this in if you have already run the previous chunk of code
trait_dat <- read.csv("data/species-trait-dat.csv")

## join rel abun and species trait
## make dataset puvot longer
rel_abun_trait <- relAbun %>%
  pivot_longer(-species, names_to = "Time Period", values_to = "Relative abundance") %>% left_join(., trait_dat, by=c("species"="scientific_name")) %>%
   filter(Habitat.type != "Aquatic") %>%
  ungroup()
  
# visualization by habitat affiliation
rel_abun_trait$Habitat.type <- factor(rel_abun_trait$Habitat.type, levels=c("Forest", "Grassland", "Generalist", "Wooded Habitat"))

fig_relAbun_habitat <- ggplot(rel_abun_trait, aes(x=`Time Period`, y=`Relative abundance`, fill = Habitat.type)) +  geom_boxplot(alpha=0.7) +  
  scale_y_log10() +
  scale_fill_scico_d(palette = "roma") + theme_bw() +
  geom_point(size=4,shape=21, alpha=0.5,
             color="black",fill="grey70") +
  geom_jitter(width=0.2,size=4,shape=21, alpha=0.5,
              color="black",fill="grey70") +
    labs(x="\nTime Period", 
       y="Relative abundance\n") + 
    theme(axis.title = element_text(family = "Century Gothic",
      size = 14, face = "bold"),
        axis.text = element_text(family="Century Gothic",size = 14),
      legend.title = element_text(family="Century Gothic",
                                    size = 14, face = "bold"),
        legend.key.size = unit(1,"cm"),
        legend.text = element_text(family="Century Gothic",size = 14))
  
ggsave(fig_relAbun_habitat, filename = "figs/fig_relAbun_habitat_landscapeLevel.png", width = 15, height = 10, device = png(), units = "in", dpi = 300)
```

Repeating the above analysis at the site-level
```{r}
# we remove modern site_code and get only distinct rows
hist_pre1900 <- hist_occ_sites_pre1900[,c(1:3)] %>%
  distinct(.)

relSpAbund_1850_1900 <- c()

for(i in 1:length(unique(hist_pre1900$historical_site_code))) {
  
  a <- hist_pre1900 %>%
  filter(historical_site_code == unique(hist_pre1900$historical_site_code)[i])

  b <- tibble::deframe(a[,c(1,3)]) ## choosing 1850-1900 data
  
  loc <- unique(hist_pre1900$historical_site_code)[i] # what's the site_code?
  
  relSpAbund_1850_1900[[i]] <- dirch_stats(b) %>%
  dplyr::select(species, bayes_mean)

  old_varname <-  c('species','bayes_mean')
  new_varname <- c('scientific_name', paste(loc,"_","1850", sep = ""))

  relSpAbund_1850_1900[[i]] <- relSpAbund_1850_1900[[i]] %>%
      data.table::setnames(old = old_varname, new = new_varname)

  names(relSpAbund_1850_1900)[i] <- loc 
}

# save object but full_join across lists
relSpAbund_1850_1900 <- relSpAbund_1850_1900 %>%
    reduce(.f=full_join)

## repeating the above across other time periods
##  time period 1900-1950

# we remove modern site_code and get only distinct rows
hist_1900_1950 <- hist_occ_sites_1900_1950[,c(1:3)] %>%
  distinct(.)

relSpAbund_1900_1950 <- c()

for(i in 1:length(unique(hist_1900_1950$historical_site_code))) {
  
  a <- hist_1900_1950 %>%
  filter(historical_site_code == unique(hist_1900_1950$historical_site_code)[i])

  b <- tibble::deframe(a[,c(1,3)]) ## choosing 1900-1950 data
  
  loc <- unique(hist_1900_1950$historical_site_code)[i] # what's the site_code?
  
  relSpAbund_1900_1950[[i]] <- dirch_stats(b) %>%
  dplyr::select(species, bayes_mean)

  old_varname <-  c('species','bayes_mean')
  new_varname <- c('scientific_name', paste(loc,"_","1900", sep = ""))

  relSpAbund_1900_1950[[i]] <- relSpAbund_1900_1950[[i]] %>%
      data.table::setnames(old = old_varname, new = new_varname)

  names(relSpAbund_1900_1950)[i] <- loc 
}

# save object but full_join across lists
relSpAbund_1900_1950 <- relSpAbund_1900_1950 %>%
    reduce(.f=full_join)

##  time period 2021
mod_2021 <- mod_occ_sites[,c(1,3,5)] %>%
  group_by(scientific_name, historical_site_code) %>%
  mutate("2021" = sum(`2021`)) %>%
  distinct(.)

relSpAbund_2021 <- c()

for(i in 1:length(unique(mod_2021$historical_site_code))) {
  
  a <- mod_2021 %>%
  filter(historical_site_code == unique(mod_2021$historical_site_code)[i])

  b <- tibble::deframe(a[,c(1,2)]) ## choosing modern data
  
  loc <- unique(mod_2021$historical_site_code)[i] # what's the site_code?
  
  relSpAbund_2021[[i]] <- dirch_stats(b) %>%
  dplyr::select(species, bayes_mean)

  old_varname <-  c('species','bayes_mean')
  new_varname <- c('scientific_name', paste(loc,"_","2021", sep = ""))

  relSpAbund_2021[[i]] <- relSpAbund_2021[[i]] %>%
      data.table::setnames(old = old_varname, new = new_varname)

  names(relSpAbund_2021)[i] <- loc 
}

# save object but full_join across lists
relSpAbund_2021 <- relSpAbund_2021 %>%
    reduce(.f=full_join)

## join all the datasets together
relSpAbund_by_site <- full_join(relSpAbund_1850_1900, relSpAbund_1900_1950, by = "scientific_name") %>%
  full_join(., relSpAbund_2021, by = "scientific_name") %>%
  arrange(scientific_name) %>%
  replace(is.na(.), 0) 
```

## Visualizing differences in relative abundance over time at the site level
```{r}
## preparing dataframe for visualization
relSpAbund <- relSpAbund_by_site %>%
  pivot_longer(!scientific_name, names_to = "site_year", values_to = "relAbundance") %>%
  separate(site_year, into = c("site", "year"), sep = "_")

## Please note: for the sake of downstream analysis, we have renamed 1850-1900 as 1850, 1900-1950 as 1900. In other words, each year - except for 2021 represents a 50 year time period. 

fig_relAbun_siteLevel <- relSpAbund %>%
  dplyr::select(-scientific_name) %>%
  # pivot_longer(everything()) %>%
  ggplot(aes(x=year, y = relAbundance,
             fill = year)) +
  geom_boxplot() +
  scale_y_log10() +
  #geom_point(size=4,shape=21, alpha=0.5,
  #           color="black",fill="grey70") +
  #geom_jitter(width=0.2,size=4,shape=21, alpha=0.5,
  #            color="black",fill="grey70") +
  facet_wrap(~site)+
  xlab("Time period") +
  ylab("Relative abundance") +
  scale_fill_scico_d(palette = "batlow") +
  theme_bw() +
  theme(axis.title = element_text(family = "Century Gothic",
      size = 14, face = "bold"),
        axis.text = element_text(family="Century Gothic",size = 14),
      legend.title = element_text(family="Century Gothic",
                                    size = 14, face = "bold"),
        legend.key.size = unit(1,"cm"),
        legend.text = element_text(family="Century Gothic",size = 14)) +
  guides(fill="none")

ggsave(fig_relAbun_siteLevel, filename = "figs/fig_relAbun_siteLevel.png", width = 15, height = 10, device = png(), units = "in", dpi = 300)
```

Takeaways: No clear differences in relative abundances over time seen at the landscape level and site-level. Due to incomplete sampling, there are some sites that show a relatively higher abundance in earlier time periods when compared to modern occurrence. Unsure if the decrease seen between 1950 to 2000 is a real pattern/artifact of incomplete sampling.
