==> 10_persistence-probabilities.Rmd <==
---
editor_options: 
  chunk_output_type: console
---

### Estimating species persistence probabilities

In this script, we will estimate probabilities of persistence for each species over time by accounting for sampling method across different years (eg. specimen collected vs. observation for historical data and heard vs. seen for modern occurrence data)

```{r}
### Load necessary libraries
library(dplyr)
library(stringr)
library(tidyverse)
library(scico)
library(RColorBrewer)
library(extrafont)
library(sf)
library(raster)
library(data.table)

# source custom functions
source("code/03_persistence-probability-functions.R")
```

### Load list of resurvey locations

We will load the list of sites across which a modern resurvey was carried out, along with elevation rasters.
```{r}
# load list of resurvey locations and add elevation as a variable
# remove the modern resurvey locations only (ie. ASFQ)
sites <- read.csv("data/list-of-resurvey-locations.csv")
sites <- sites[-c(1:50),]

# convert to sf object and transform
sites <- st_as_sf(sites, coords = c("longitude", "latitude")) %>%
  `st_crs<-`(4326) %>%
  st_transform(32643)

# add elevation raster
alt <- raster("data/spatial/elevation/alt") # this layer is not added to github as a result of its large size and can be downloaded from SRTM (Farr et al. (2007))

# extract values from that raster (note: transformation of coordinate system)
elev <- extract(alt, sites)
sites <- cbind(sites, elev)
```

### Load in historical occurrence data
```{r}
## read in historical occurrence data
hist_occ <- read.csv("data/historical-occurrence-data.csv")

# We add in two columns to the above dataframe:
# pci_lower and pci_upper
# here pci is a metric from Thompson et al. 2017 which essentially translates to the probability that the taxon is correctly identified as extant. 

# We make a distinction between the upper and lower probabilities that are assigned to the above metric. 
# We follow Lees et al. (2021) and Thompson et al. (2017) who assigned values of 0.95-0.99 for specimen based records/museum records and conservative values of 0.5-0.65 for journal data (often reporting observations/sightings and not specimens collected). Such probabilities are assigned to account for varying efforts that cannot be quantified from historical observations. 

hist_occ <-  hist_occ %>%
  mutate("pci_lower" = case_when(institutionCode == "JournalData" ~ 0.5,
                                 TRUE ~ 0.95)) %>%
  mutate("pci_upper" = case_when(institutionCode == "JournalData" ~ 0.65,
                                 TRUE ~ 0.99))

### fix the count column
hist_occ <-  hist_occ %>%
  mutate(histCount = case_when(individualCount == "X" ~ 1,
                               individualCount == "" ~ 1,
                               TRUE ~ as.numeric(individualCount)))
```


### Load in modern occurrence data
```{r}
## read in modern occurrence data
mod_occ <- read.csv("data/modern-occurrence-data.csv")

## remove sites without historical_site_code
mod_occ <- mod_occ %>%
  filter(!is.na(historical_site_code))

## add the pci_lower and pci_upper values
## here, we do not make a distinction between heard, seen and flyover observations and assign a pci_lower of 0.95 and a pci_upper of 0.99

mod_occ <- mod_occ %>%
  mutate("pci_lower" =  0.95) %>%
  mutate("pci_upper" = 0.99)
```

## Combine modern and historical occurrence data

Choose the columns necessary for further calculations of probabilities of persistence
```{r}
## historical data
hist_probDat <-  hist_occ %>%
  dplyr::select(scientific_name, year, historical_site_code, pci_lower,
         pci_upper, histCount) %>%
  group_by(scientific_name, year, historical_site_code) %>%
  mutate(histCount = sum(histCount)) %>%
  rename(., abundance = "histCount") %>%
  distinct(.) %>%
  ungroup()

## modern data
mod_probDat <- mod_occ %>%
  mutate(year = 2021) %>% # change year to 2021 for ease of modeling
  rename(., abundance = "number") %>%
  dplyr::select(scientific_name, year, historical_site_code, pci_lower,
         pci_upper, abundance) %>%
  group_by(scientific_name, year, historical_site_code) %>%
  mutate(abundance = sum(abundance)) %>%
  distinct(.) %>%
  filter(!is.na(scientific_name)) %>%
  ungroup()
```

## Calculate probabilities of persistence

We use functions from Thompson et al. 2017 to calculate the same. 

The analysis needs to be done at the site level and at the landscape level
```{r}
# join the above two datasets for ease of calculation
probDat <- full_join(hist_probDat, mod_probDat)

# note: unlike example datasets, we have data for the same year from different locations for the same species. 

## modeling data by species at the site level

# years in which a species was not detected at all, we insert default probabilities for eps, pi and pr - where eps is the the proportion of the taxon's habitat within its likely entire range that was surveyed (0 ≤ ε ≤ 1).
# p(i) = the probability that the taxon, or recent evidence of it, could have been reliably identified in the survey if it had been recorded.
# p(r) = the probability that the taxon, or recent evidence of it, would have been recorded in the survey.

eps.passive = c(0,	0.05)
pi.passive = c(0.10,	0.65)
pr.passive = c(0.40,	0.60)
pas.sur = c(eps.passive ,pi.passive,pr.passive)

# please refer to Thompson et al. 2017 for more details/information

# run the loop to create a list of plots
allPlots <- list()

# save the persistence prob in a dataframe for later use
spec_site_persProb <- data.frame()

for(i in 1:length(unique(probDat$scientific_name))) {
  
  specDat <- probDat %>%
    filter(scientific_name == unique(probDat$scientific_name)[i])
  sciName <- unique(probDat$scientific_name)[i]
  
  sitePlots <- list()
  
  for(j in 1:length(unique(specDat$historical_site_code))) {
    
  recordings <- specDat %>%
    filter(historical_site_code == unique(specDat$historical_site_code)[j])
    siteName <- unique(specDat$historical_site_code)[j]
  
  recordings <- as.data.frame(recordings[,c(2,4,5)]) # select only year & pci
    
    # check years are unique - one record per year
    length(unique(recordings$year)) == nrow(recordings)

    # total number of years (including years without data)
    years <- seq(min(recordings[,'year']),2030 ,by = 1)
    
    # find years in which no data was recorded
    pas_sur_years <- years[!years %in% recordings[,1]]
    pas_sur_years <- cbind(pas_sur_years, t(replicate(length(pas_sur_years), pas.sur)))
    surveys <- as.data.frame(pas_sur_years)
    names(surveys) <- c("year",	"eps_lower","eps_upper",
                        "pi_lower","pi_upper", "pr_lower",
                        "pr_upper")
    surveys <- surveys[order(surveys[,"year"]),]
    rownames(surveys) <- NULL
    
    # create an object that adds a 1 and 0 for each record/no record
    rec.year <- cbind(recordings[,'year'],1) 
    rec.year = rbind(rec.year,cbind(surveys[,'year'],0)) 
    rec.year = rec.year[order(rec.year[,1]),]
    
     # some more tests
    nrow(rec.year) == (nrow(recordings) + nrow(surveys))
    (nrow(recordings) + nrow(surveys)) == length(years)
    
    # calculate probability of persistence and 95% intervals
    PXt <- px.mid()
    
    # visualizing probabilities of persistence by species and site
    specBySiteDat <- as.data.frame(cbind(years,PXt,sciName,siteName))
    colnames(specBySiteDat) <- c("years", "sd_lwr", "PXt", "sd_upr", "PXt_min","PXt_max","scientific_name","historical_site_code")
    
    # store it in a dataframe
    spec_site_persProb <- rbind(specBySiteDat, spec_site_persProb)
    
    # plots
    sitePlots[[j]] <- ggplot(specBySiteDat, aes(x = years, y = PXt))+
      geom_ribbon(aes(ymin=sd_lwr, ymax = sd_upr), linetype = 0, alpha = 0.4)+
  geom_ribbon(aes(ymin=PXt_min, ymax = PXt_max), linetype = 0, alpha = 0.2) +
  scale_color_manual(values = c("black", "darkred")) +
  scale_fill_manual(aesthetics = "fill", values = c("grey10", "grey50")) + 
  geom_line(size = 1) +
  ylim(c(0,1)) +
  scale_x_continuous(breaks = seq(min(years), max(years),10)) +
  ylab("Probability that taxon is extant")+
  theme_bw() +
  ggtitle(paste(sciName,"-",siteName, sep="")) +
  labs(
    x = "\nYears",
    y = "Probability that a taxon is extant\n"
  ) +
  theme(
    plot.title = element_text(family = "Century Gothic", face="bold.italic"),
    axis.title = element_text(
      family = "Century Gothic",
      size = 14, face = "bold"
    ),
    axis.text = element_text(family = "Century Gothic", size = 13),
    legend.title = element_text(family = "Century Gothic"),
    legend.text = element_text(family = "Century Gothic")) + 
    guides(size= "none",
           fill = "none")
  }
  allPlots <- c(allPlots, sitePlots)
  rm(sitePlots)
}
  
# plot and save as a single pdf
options(max.print = 2000)
cairo_pdf(
  filename = "figs/fig_species_by_site_persistenceProb.pdf",
  onefile = TRUE,
  width = 12,
  height = 7
)
allPlots
dev.off()
```


### Rerunning the above analysis but at the landscape level for each species

First, we will prep the historical and modern occurrence data
```{r}
## historical data
hist_probDat <-  hist_occ %>%
  dplyr::select(scientific_name, year, pci_lower,
         pci_upper, histCount) %>%
  group_by(scientific_name, year) %>%
  mutate(histCount = sum(histCount)) %>%
  rename(., abundance = "histCount") %>%
  distinct(.) %>%
  ungroup()

## modern data
mod_probDat <- mod_occ %>%
  mutate(year = 2021) %>% # change year to 2021 for ease of modeling
  rename(., abundance = "number") %>%
  dplyr::select(scientific_name, year, pci_lower,
         pci_upper, abundance) %>%
  group_by(scientific_name, year) %>%
  mutate(abundance = sum(abundance)) %>%
  distinct(.) %>%
  filter(!is.na(scientific_name)) %>%
  ungroup()
```


## Calculate probabilities of persistence

We use functions from Thompson et al. 2017 to calculate the same. 

The analysis is now being done at the landscape level
```{r}
# join the above two datasets for ease of calculation
probDat <- full_join(hist_probDat, mod_probDat)

## modeling data by species at the landscape level

# years in which a species was not detected at all, we insert default probabilities for eps, pi and pr - where eps is the the proportion of the taxon's habitat within its likely entire range that was surveyed (0 ≤ ε ≤ 1).
# p(i) = the probability that the taxon, or recent evidence of it, could have been reliably identified in the survey if it had been recorded.
# p(r) = the probability that the taxon, or recent evidence of it, would have been recorded in the survey.

eps.passive = c(0,	0.05)
pi.passive = c(0.10,	0.65)
pr.passive = c(0.40,	0.60)
pas.sur = c(eps.passive ,pi.passive,pr.passive)

# please refer to Thompson et al. 2017 for more details/information

# run the loop to create a list of plots
allPlots <- list()

# save the persistence prob in a dataframe for later use
spec_persProb <- data.frame()

for(i in 1:length(unique(probDat$scientific_name))) {
  
  recordings <- probDat %>%
  filter(scientific_name == unique(probDat$scientific_name)[i])
  sciName <- unique(probDat$scientific_name)[i]
  
  recordings <- as.data.frame(recordings[,c(2,3,4)]) # select only year & pci
    
    # check years are unique - one record per year
    length(unique(recordings$year)) == nrow(recordings)

    # total number of years (including years without data)
    years <- seq(min(recordings[,'year']),2030 ,by = 1)
    
    # find years in which no data was recorded
    pas_sur_years <- years[!years %in% recordings[,1]]
    pas_sur_years <- cbind(pas_sur_years, t(replicate(length(pas_sur_years), pas.sur)))
    surveys <- as.data.frame(pas_sur_years)
    names(surveys) <- c("year",	"eps_lower","eps_upper",
                        "pi_lower","pi_upper", "pr_lower",
                        "pr_upper")
    surveys <- surveys[order(surveys[,"year"]),]
    rownames(surveys) <- NULL
    
    # create an object that adds a 1 and 0 for each record/no record
    rec.year <- cbind(recordings[,'year'],1) 
    rec.year = rbind(rec.year,cbind(surveys[,'year'],0)) 
    rec.year = rec.year[order(rec.year[,1]),]
    
     # some more tests
    nrow(rec.year) == (nrow(recordings) + nrow(surveys))
    (nrow(recordings) + nrow(surveys)) == length(years)
    
    # calculate probability of persistence and 95% intervals
    PXt <- px.mid()
    
    # visualizing probabilities of persistence by species and site
    specDat <- as.data.frame(cbind(years,PXt,sciName))
    colnames(specDat) <- c("years", "sd_lwr", "PXt", "sd_upr", "PXt_min","PXt_max","sciName")
    
    # store it in a dataframe
    spec_persProb <- rbind(specDat, spec_persProb)
    
    # plots
    allPlots[[i]] <- ggplot(specDat, aes(x = years, y = PXt))+
      geom_ribbon(aes(ymin=sd_lwr, ymax = sd_upr), linetype = 0, alpha = 0.4)+
  geom_ribbon(aes(ymin=PXt_min, ymax = PXt_max), linetype = 0, alpha = 0.2) +
  scale_color_manual(values = c("black", "darkred")) +
  scale_fill_manual(aesthetics = "fill", values = c("grey10", "grey50")) + 
  geom_line(size = 1) +
  ylim(c(0,1)) +
  scale_x_continuous(breaks = seq(min(years), max(years),10)) +
  ylab("Probability that taxon is extant")+
  theme_bw() +
  ggtitle(sciName) +
  labs(
    x = "\nYears",
    y = "Probability that a taxon is extant\n"
  ) +
  theme(
    plot.title = element_text(family = "Century Gothic", face="bold.italic"),
    axis.title = element_text(
      family = "Century Gothic",
      size = 14, face = "bold"
    ),
    axis.text = element_text(family = "Century Gothic", size = 13),
    legend.title = element_text(family = "Century Gothic"),
    legend.text = element_text(family = "Century Gothic")) + 
    guides(size= "none",
           fill = "none")
  }
  
# plot and save as a single pdf
cairo_pdf(
  filename = "figs/fig_species_persistenceProb_landscapeLevel.pdf",
  onefile = TRUE,
  width = 12,
  height = 7
)
allPlots
dev.off()

## save dataframes to file
write.csv(spec_site_persProb, "data/species-by-site-persProb.csv", row.names=F)
write.csv(spec_persProb, "data/species-persProb-landscapeLevel.csv", row.names=F)
```

## Load the species persistence probabilities at the site level

We choose data for min and max years to see how historic and modern persistence probabilities vary
```{r}
## select data from the min and max years
min_siteDat <- spec_site_persProb %>%
  dplyr::select(years,PXt,scientific_name,historical_site_code) %>%
  group_by(scientific_name, historical_site_code) %>%
  slice(which.min(years)) %>%
  mutate(pers_prob_year = "minYear") %>%
  ungroup()

max_siteDat <- spec_site_persProb %>%
  dplyr::select(years,PXt,scientific_name,historical_site_code) %>%
  group_by(scientific_name, historical_site_code) %>%
  filter(years== 2021) %>%
  mutate(pers_prob_year = "maxYear") %>%
  ungroup()

# join the above two datasets together
siteDat <- bind_rows(min_siteDat, max_siteDat)

# visualization of persistence probabilities between min and max year by site

siteDat$pers_prob_year <- factor(siteDat$pers_prob_year, levels = c("minYear", "maxYear"))
siteDat$PXt <- as.numeric(siteDat$PXt)

fig_site_persProb <- ggplot(siteDat, aes(x=pers_prob_year, y=PXt, fill=pers_prob_year)) +  geom_boxplot(alpha=0.7) +  
  scale_fill_scico_d(palette = "roma") +
    theme_bw() +
    labs(x="\nHistorical vs. modern persistence probability", 
       y="Persistence probabilty\n") +
  facet_wrap(~historical_site_code)+
  scale_x_discrete(labels = c('Historic','Modern')) +
    theme(axis.title = element_text(family = "Century Gothic",
      size = 14, face = "bold"),
        axis.text = element_text(family="Century Gothic",size = 14),
        legend.position = "none")

ggsave(fig_site_persProb, filename = "figs/fig_persProb_siteLevel.png", width = 15, height = 10, device = png(), units = "in", dpi = 300)
```

## Load species persistence probabilities at the landscape level

We choose persistence probabilities for min and max years and repeat analysis we carried out previously. 
```{r}
## select data from the min and max years
min_siteDat <- spec_persProb %>%
  dplyr::select(years,PXt,sciName) %>%
  group_by(sciName) %>%
  slice(which.min(years)) %>%
  mutate(pers_prob_year = "minYear") %>%
  ungroup()

max_siteDat <- spec_persProb %>%
  dplyr::select(years,PXt,sciName) %>%
  group_by(sciName) %>%
  filter(years== 2021) %>%
  mutate(pers_prob_year = "maxYear") %>%
  ungroup()

# join the above two datasets together
specDat <- bind_rows(min_siteDat, max_siteDat)

# visualization of persistence probabilities between min and max year by site

specDat$pers_prob_year <- factor(specDat$pers_prob_year, levels = c("minYear", "maxYear"))
specDat$PXt <- as.numeric(specDat$PXt)

fig_persProb_landscapeLevel <- ggplot(specDat, aes(x=pers_prob_year, y=PXt, fill=pers_prob_year)) +  geom_boxplot(alpha=0.7) +  
  scale_fill_scico_d(palette = "roma") +
    theme_bw() +
    labs(x="\nHistorical vs. modern persistence probability", 
       y="Persistence probabilty\n") +
  scale_x_discrete(labels = c('Historic','Modern')) +
    theme(axis.title = element_text(family = "Century Gothic",
      size = 14, face = "bold"),
        axis.text = element_text(family="Century Gothic",size = 14),
        legend.position = "none")

ggsave(fig_persProb_landscapeLevel, filename = "figs/fig_persProb_landscapeLevel.png", width = 15, height = 10, device = png(), units = "in", dpi = 300)
```

## Summarizing persistence probabilities by species trait data

Now that we have persistence probabilities over time, we can examine how they vary for species by site and at the landscape level over time for habitat generalists and specialists. 

#### Let's first curate the species trait data
```{r}
## load state of India's birds data to get habitat affiliation
species_codes <- read.csv("data/species-codes.csv")
habitat_SoIB <- read.csv("data/SoIB-data.csv")

# join the two dataframes
habitat_dat <- left_join(species_codes, habitat_SoIB, by=c("scientific_name"="scientific_name"))

# write the above to data folder and recode habitat affiliation
# write.csv(habitat_dat, "data/species-trait-dat.csv", row.names = F)

# note: the classification of habitat type was taken from the 2020 State of India's Birds report which had a comprehensive list of habitat types for each species. We classified species as forest associated even if it preferred some level of woodland habitat/shrubland (including throny scrubland and dry deciduous forests - lower elevations of the Nilgiris for instance), generalist - if it was found near settlements and habitations and grassland - if it occupied/preferred dry/wet/hill grassland/scrubland or even marshes (exception here: eurasian woodcock)

## load the species trait data (that includes habitat) and merge with the AVONET dataset
## please note that it is being loaded after curating and adding a column for habitat type based on our understanding of species natural history
hab_dat <- read.csv("data/species-habitat-dat.csv")
avonet <- read.csv("data/AVONET-data.csv")

# join the above two datasets
trait_dat <- left_join(hab_dat, avonet[,c(1,10:20,29:31)], by=c("scientific_name"="Species2"))

## Write the trait data to file
## write.csv(trait_dat,"data/species-trait-dat.csv",row.names = F)
```

## Examining persistence probabilities at the site level in conjunction with species traits
```{r}
## join pers prob and trait datasets
site_dat_trait <- left_join(spec_site_persProb,trait_dat, by=c("scientific_name"="scientific_name"))

## select data from the min and max years
min_siteDat_trait <- site_dat_trait %>%
  group_by(scientific_name, historical_site_code) %>%
  slice(which.min(years)) %>%
  mutate(pers_prob_year = "minYear") %>%
  filter(Habitat.type != "Aquatic") %>%
  ungroup()

max_siteDat_trait <- site_dat_trait %>%
  group_by(scientific_name, historical_site_code) %>%
  filter(years== 2021) %>%
  mutate(pers_prob_year = "maxYear") %>%
  filter(Habitat.type != "Aquatic") %>%
  ungroup()

# join the above two datasets together
siteDatTrait <- bind_rows(min_siteDat_trait, max_siteDat_trait)

# visualization of persistence probabilities between min and max year by site

siteDatTrait$pers_prob_year <- factor(siteDatTrait$pers_prob_year, levels = c("minYear", "maxYear"))
siteDatTrait$PXt <- as.numeric(siteDatTrait$PXt)

## examining habitat type at site level
fig_site_habitat <- ggplot(siteDatTrait, aes(x=pers_prob_year, y=PXt, fill=Habitat.type)) +  geom_boxplot(alpha=0.7) +  
  scale_fill_scico_d(palette = "roma",
                     labels=c("Generalist","Forest","Grassland")) +
    theme_bw() +
    labs(x="\nHistorical vs. modern persistence probability", 
       y="Persistence probabilty\n") +
  facet_wrap(~historical_site_code)+
  scale_x_discrete(labels = c('Historic','Modern')) +
    theme(axis.title = element_text(family = "Century Gothic",
      size = 14, face = "bold"),
        axis.text = element_text(family="Century Gothic",size = 14),
      legend.title = element_text(family="Century Gothic",
                                    size = 14, face = "bold"),
        legend.key.size = unit(1,"cm"),
        legend.text = element_text(family="Century Gothic",size = 14))

ggsave(fig_site_habitat, filename = "figs/fig_persProb_siteLevel_habitatType.png", width = 15, height = 10, device = png(), units = "in", dpi = 300)
```

## Examining persistence probabilities at the landscape level in conjunction with species traits
```{r}
## join pers prob and trait datasets
spec_dat_trait <- left_join(spec_persProb,trait_dat, by=c("sciName"="scientific_name"))

## select data from the min and max years
min_specDat_trait <- spec_dat_trait %>%
  group_by(sciName) %>%
  slice(which.min(years)) %>%
  mutate(pers_prob_year = "minYear") %>%
  filter(Habitat.type != "Aquatic") %>%
  ungroup()

max_specDat_trait <- spec_dat_trait %>%
  group_by(sciName) %>%
  filter(years== 2021) %>%
  mutate(pers_prob_year = "maxYear") %>%
  filter(Habitat.type != "Aquatic") %>%
  ungroup()

# join the above two datasets together
specDatTrait <- bind_rows(min_specDat_trait, max_specDat_trait)

# visualization of persistence probabilities between min and max year by site

specDatTrait$pers_prob_year <- factor(specDatTrait$pers_prob_year, levels = c("minYear", "maxYear"))
specDatTrait$PXt <- as.numeric(specDatTrait$PXt)

## examining habitat type at site level
fig_spec_habitat <- ggplot(specDatTrait, aes(x=pers_prob_year, y=PXt, fill=Habitat.type)) +  geom_boxplot(alpha=0.7) +  
  scale_fill_scico_d(palette = "roma",
                     labels=c("Generalist","Forest","Grassland")) +
    theme_bw() +
    labs(x="\nHistorical vs. modern persistence probability", 
       y="Persistence probabilty\n") +
  scale_x_discrete(labels = c('Historic','Modern')) +
    theme(axis.title = element_text(family = "Century Gothic",
      size = 14, face = "bold"),
        axis.text = element_text(family="Century Gothic",size = 14),
      legend.title = element_text(family="Century Gothic",
                                    size = 14, face = "bold"),
        legend.key.size = unit(1,"cm"),
        legend.text = element_text(family="Century Gothic",size = 14))

ggsave(fig_spec_habitat, filename = "figs/fig_persProb_landscapeLevel_habitatType.png", width = 15, height = 10, device = png(), units = "in", dpi = 300)
```
