==> 14_species-turnover-over-time.Rmd <==
---
editor_options: 
  chunk_output_type: console
---

## Estimating species turnover across time periods

We use temporal diversity metrics from Hallett et al. 2016 based on Collins et al. 2008 and Cleland et al. 2013. We use functions from the package codyn to calculate temporal diversity indices. Please see this link for a detailed vignette: <https://cran.r-project.org/web/packages/codyn/vignettes/Temporal_Diversity_Indices.html>

### Load necessary libraries

```{r}
library(dplyr)
library(stringr)
library(readr)
library(scico)
library(RColorBrewer)
library(extrafont)
library(sf)
library(raster)
library(data.table)
library(codyn)
```

### Load list of resurvey locations

We will load the list of sites across which a modern resurvey was carried out, along with elevation rasters.

```{r}
# load list of resurvey locations and add elevation as a variable
# remove the modern resurvey locations only (ie. ASFQ)
sites <- read.csv("data/list-of-resurvey-locations.csv")
sites <- sites[-c(1:50),]

# convert to sf object and transform
sites <- st_as_sf(sites, coords = c("longitude", "latitude")) %>%
  `st_crs<-`(4326) %>%
  st_transform(32643)

# add elevation raster
alt <- raster("data/spatial/elevation/alt") # this layer is not added to github as a result of its large size and can be downloaded from SRTM (Farr et al. (2007))

# extract values from that raster (note: transformation of coordinate system)
elev <- extract(alt, sites)
sites <- cbind(sites, elev)
```

For the historical occurrence data, we will choose data from three time periods: 1850-1900, 1900-1950 and 1950-2000.

```{r}
## read in historical occurrence data
hist_occ <- read_csv("data/historical-occurrence-data.csv")

### prep data and bin by time periods
hist_occ <- hist_occ %>%
  mutate(
    year_bin = case_when(
      year <= 1900 ~ "1850-1900",
      between(year, 1901, 1950) ~ "1900-1950",
      between(year, 1951, 2000) ~ "1950-2000",
      between(year, 2001, 2020) ~ "2021"
    ),
    histCount = as.numeric(individualCount),
    histCount = if_else(is.na(histCount), 1, histCount)
  ) %>%
  dplyr::select(
    scientific_name, year, year_bin, historical_site_code, histCount
  ) %>%
  group_by(scientific_name, year, year_bin, historical_site_code) %>%
  summarise(histCount = sum(histCount, na.rm = TRUE)) %>%
  ungroup()

## read in modern occurrence data
mod_occ <- read_csv("data/modern-occurrence-data.csv")

## remove sites without historical_site_code
mod_occ <- mod_occ %>%
  filter(!is.na(historical_site_code))

## for the sake of downstream analysis, change the year column
mod_occ$year <- 2021

## subset data
mod_occ <- mod_occ %>%
  group_by(scientific_name, year, historical_site_code) %>%
  summarise(histCount = sum(number)) %>% # rename to histCount
  ungroup() %>%
  mutate(
    year_bin = "2021"
  )
```

## Preparing a dataframe to calculate species turnover

```{r}
# join all data
occ = full_join(hist_occ, mod_occ)

# count number of years per site, remove 
occ = occ |> 
  group_by(
    historical_site_code
  ) |> 
  mutate(
    unique_years = length(unique(year))
  ) |> 
  filter(
   unique_years > 1 
  ) |> 
  dplyr::select(-unique_years)
```

**Two modern sites sampled only in 2021 are lost! These are CNRG and FHEK.**

## Get turnover for each site separately

```{r}
# calculate turnover for a single site
occ |> 
  filter(historical_site_code == "CNRG", !is.na(scientific_name)) |> 
  arrange(year) |> View()

# observations by site
occ_site = split(x = occ, f = occ$historical_site_code)

# apply function turnover over sites
turnover_site = lapply(
  occ_site, function(df) {
    
    message(sprintf("this is site: %s", unique(df$historical_site_code)))
    
    codyn::turnover(
      df = df,
      time.var = "year",
      species.var = "scientific_name",
      abundance.var = "histCount",
      replicate.var = "historical_site_code"
    )
  }
)
```

## Get turnover for all sites together

```{r}
# code for turnover at all sites together --- appears to be the same as code above
codyn::turnover(
  df = occ,
  time.var = "year",
  species.var = "scientific_name",
  abundance.var = "histCount",
  replicate.var = "historical_site_code"
)
```

## Rank shifts

```{r}
# rank shift calculation works without single year sites
codyn::rank_shift(
  df = occ,
  time.var = "year",
  species.var = "scientific_name",
  abundance.var = "histCount",
  replicate.var = "historical_site_code"
)
```
