==> 15_species-persProb-relAbun-envtChange.Rmd <==
---
output: html_document
editor_options: 
  chunk_output_type: console
---

# Species persistence probabilities and relative abundance as a function of environmental changes

In this script, we will model the effect of species persistence probabilities and relative abundance as a function of environmental changes at each of the sampling sites (environmental changes are calculated as differences in values of a predictor between modern vs. historical time period)

In the first half of the script, we will prepare the environmental predictors at each of the sampling sites

## Load necessary libraries
```{r}
# load libs
library(sf)
library(readr)
library(dplyr)
library(terra)
library(tidyr)
library(mapview)
library(sjPlot)
library(glmmTMB)
library(ggplot2)
library(effects)
library(lme4)

# function to z-transform data
scale_z <- function(x){
  (x - mean(x, na.rm=TRUE))/sd(x, na.rm=TRUE)
}
```

## Load land cover data from 1870 and 2017
```{r}
# load 1870 and 2017 shp
lc_1870 <- st_read("data/spatial/1870.shp")
lc_2017 <- st_read("data/spatial/2017.shp")

# Prior to further analysis, the classes have to be renamed so that they match classifications done in 1870 and 2017
# For the sake of further downstream analysis, we ensure that the proportion of land cover estimation is consistent with class names
# We are unable to distinguish between Tea and Timber plantations in the 1870 map and hence, we club Timber and Tea plantations in 2017 as a single class (Plantations)

# cleaning up the 1870 map
lc_1870$class[lc_1870$class=="cultivated tracts"] <- "Agricultural Land"
lc_1870$class[lc_1870$class=="plantations"] <- "Plantations (Tea & Timber)"
lc_1870$class[lc_1870$class=="settlements"] <- "Settlements"
lc_1870$class[lc_1870$class=="waterbodies"] <- "Water Bodies"

# combining classes in the 2017 map
lc_2017$class[lc_2017$class=="Timber Plantations"] <- "Plantations (Tea & Timber)"
lc_2017$class[lc_2017$class=="Tea Plantations"] <- "Plantations (Tea & Timber)"
lc_2017$class[lc_2017$class=="Water bodies"] <- "Water Bodies"
```

## Load modern resruvey locations
```{r}
# load sites
sites <- read_csv("data/list-of-resurvey-locations.csv")
sites <- distinct(sites, modern_site_code, historical_site_code, longitude, latitude)
sites <- sites[-c(1:50),]
```

## Create spatial data within a 250m buffer for each site
```{r}
# make spatial data from sites, with a 250m buffer
sites_sf = st_as_sf(
  sites, coords = c("longitude", "latitude"),
  crs = 4326
) |> 
  st_transform(32643) |> 
  st_buffer(250)
```

```{r}
# get intersection with 1870 and 2017 data
lc_sites <- lapply(
  list(lc_1870, lc_2017), function(x) {
    st_intersection(x, sites_sf)
  }
)

# calculate area for each lc type for sites based on modern site code
lc_sites <- lapply(lc_sites, function(df) {
  df |> 
    mutate(
      area = as.numeric(st_area(df))
    ) |> 
    st_drop_geometry() |> 
    group_by(modern_site_code, class) |> 
    summarise(prop_lc = area / sum(area))
})

# add year and merge
lc_sites <- Map(
  lc_sites, c(1870, 2017),
  f = function(df, y) {
    df$year = y
    df
  }
) |> 
  bind_rows()
```

```{r}
# get lc change sites
lc_change_dat <- lc_sites |> 
  pivot_wider(
    id_cols = c("modern_site_code", "class"),
    names_from = c("year"),
    values_from = "prop_lc",
    values_fn = mean
  ) |> 
  rename(
    year_1870 = `1870`,
    year_2017 = `2017`
  ) %>%
  left_join(., sites, by="modern_site_code") %>%
  replace(is.na(.), 0) %>%
  mutate("lc_change" = year_2017 - year_1870) #  -1 implies loss and + 1 implies gain
```

```{r}
# save as is for further processing
write_csv(
  lc_change_dat, file = "results/lc_change.csv"
)
```

## Climate change at sampling sites
```{r}
# load climate data
raster_temp = list.files("data/climate/", pattern = "t_mean", full.names = T) |> 
  lapply(rast)

raster_rain = list.files("data/climate/", pattern = "ppt", full.names = T) |> 
  lapply(rast)
```

```{r}
# get difference between first and last files
climate_change = lapply(
  list(raster_temp, raster_rain),
  function(le) {
    lapply(le, function(le2) {
      le2[[length(names(le2))]] - le2[[1]]
    })
  }
) |> 
  Reduce(f = c) # convert to single list

# get raster names
climate_change_names = lapply(climate_change, function(ra) {
  lapply(ra, function(r) {
    r |> names() |> stringr::str_extract("(.*?)(?=_\\d+)")
  })
}) |> unlist()
```

```{r}
# get mean change within 250m buffer of resurvey locs
climate_change_sites = lapply(
  climate_change, function(ra) {
    terra::extract(
      ra, 
      y = terra::vect(
        sites_sf |> 
          st_transform(4326)
      ), fun = mean
    )  
  }
)

# combine all metrics
climate_change_sites = Reduce(f = left_join, x = climate_change_sites)

# attach to sampling sites
sites <- mutate(sites, ID = seq(nrow(sites)))
clim_change_dat <- left_join(sites, climate_change_sites)
```

```{r}
# save site data
write_csv(
  clim_change_dat, file = "results/climate_change.csv"
)
```

### Load species persistence probabilities at site and landscape level
```{r}
spec_site_persProb <- read.csv( "results/species-persProb-siteLevel.csv")
spec_persProb <- read.csv("results/species-persProb-landscapeLevel.csv")
```

## Load species relative abundance
```{r}
relAbun <- read.csv("results/species-relative-abundance-siteLevel.csv")
```

## Load species trait data
```{r}
trait_dat <- read.csv("data/species-trait-dat.csv")
```

## Run mixed-effects models examining persistence probabilties and land cover change
```{r}
## change in persistence prob over time (site level)
persProb_min <- spec_site_persProb %>%
  dplyr::select(years,PXt,scientific_name, historical_site_code) %>%
  group_by(scientific_name, historical_site_code) %>%
  slice(which.min(years)) %>%
  rename(., PXt_min = "PXt") %>%
  ungroup()

persProb_max <- spec_site_persProb %>%
  dplyr::select(years,PXt,scientific_name, historical_site_code) %>%
  group_by(scientific_name, historical_site_code) %>%
  slice(which.max(years)) %>%
  rename(., PXt_max = "PXt") %>%
  ungroup()

## join the two dataframes
persProb_change <- full_join(persProb_min[,-1], persProb_max[,-1]) %>%
  mutate("persProbChange" = PXt_min - PXt_max) %>%
  ungroup()

### make land cover data wider
lc_change_glmm <- lc_change_dat %>%
  pivot_wider(names_from = class, values_from = lc_change) %>%
  ungroup() %>%
  dplyr::select(`historical_site_code`, `Shola Grassland`, `Settlements`, `Shola Forest`,`Plantations (Tea & Timber)`, `Agricultural Land`) %>%
  distinct(.) %>%
  replace(is.na(.), 0) %>% 
  group_by(historical_site_code) %>%
  summarise(across(everything(), sum)) %>%
  mutate("Shola Grassland" = case_when(`Shola Grassland` < -1 ~ -1,
                                       `Shola Grassland` > 1 ~ 1,
                                       TRUE ~ as.numeric(`Shola Grassland`))) %>%
  mutate("Settlements" = case_when(`Settlements` < -1 ~ -1,
                                       `Settlements` > 1 ~ 1,
                                       TRUE ~ as.numeric(`Settlements`))) %>%
  mutate("Shola Forest" = case_when(`Shola Forest` < -1 ~ -1,
                                       `Shola Forest` > 1 ~ 1,
                                       TRUE ~ as.numeric(`Shola Forest`))) %>%
  mutate("Plantations (Tea & Timber)" = case_when(`Plantations (Tea & Timber)`< -1 ~ -1, `Plantations (Tea & Timber)` > 1 ~ 1, TRUE ~ as.numeric(`Plantations (Tea & Timber)`))) %>%
  mutate("Agricultural Land" = case_when(`Agricultural Land` < -1 ~ -1,
                                       `Agricultural Land` > 1 ~ 1,
                                       TRUE ~ as.numeric(`Agricultural Land`))) %>%
  ungroup()

# in the above dataframe, we lost KJPN and KULR as it's a low elevation site and our land cover data does not sample that area
# hence we filter those sites from the persProbChange for the landscape change analysis

persProb_landCover <- persProb_change %>%
  filter(historical_site_code != "KJPN") %>%
  filter(historical_site_code != "KULR")
  
glmm_persProb <- left_join(persProb_landCover, lc_change_glmm, by="historical_site_code") %>% as.data.frame()

## run the glmm
glmm_lc_persProb <- glmmTMB(persProbChange ~ `Shola Grassland` + `Shola Forest` + `Settlements`+ `Plantations (Tea & Timber)` + `Agricultural Land` + (1|scientific_name) + (1|historical_site_code), 
                        data = glmm_persProb, family = beta_family(link = "logit"))

summary(glmm_lc_persProb)

# marginal effect of increase of settlements on change in persistence probability
```
Family: beta  ( logit )
Formula:          
persProbChange ~ `Shola Grassland` + `Shola Forest` + Settlements +  
    `Plantations (Tea & Timber)` + `Agricultural Land` + (1 |  
    scientific_name) + (1 | historical_site_code)
Data: glmm_persProb

     AIC      BIC   logLik deviance df.resid 
 -2279.7  -2234.3   1148.9  -2297.7     1139 

Random effects:

Conditional model:
 Groups               Name        Variance Std.Dev.
 scientific_name      (Intercept) 0.59781  0.7732  
 historical_site_code (Intercept) 0.05833  0.2415  
Number of obs: 1148, groups:  
scientific_name, 189; historical_site_code, 26

Dispersion parameter for beta family (): 12.1 

Conditional model:
                             Estimate Std. Error z value Pr(>|z|)   
(Intercept)                  -1.99310    0.70228  -2.838  0.00454 **
`Shola Grassland`            -0.44962    0.68193  -0.659  0.50968   
`Shola Forest`               -0.01850    0.09850  -0.188  0.85106   
Settlements                   0.34084    0.10389   3.281  0.00103 **
`Plantations (Tea & Timber)`  0.11122    0.16393   0.678  0.49747   
`Agricultural Land`           0.08882    0.14688   0.605  0.54536   
---
Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1


## Run the mixed-effects models, but incorporate species habitat affiliation
```{r}
glmm_persProb_habitat <- left_join(glmm_persProb, trait_dat, by = "scientific_name") %>%
  as.data.frame()

## run the glmm
## grassland birds
glmm_persProb_grassland <-  glmm_persProb_habitat %>%
  filter(Habitat.type == "Grassland")

glmm_grassland <- glmmTMB(persProbChange ~ `Shola Grassland` + `Shola Forest` + `Settlements`+ `Plantations (Tea & Timber)` + `Agricultural Land` + (1|scientific_name) + (1|historical_site_code), 
                        data = glmm_persProb_grassland, family = beta_family(link = "logit"))

summary(glmm_grassland)
```
Family: beta  ( logit )
Formula:          
persProbChange ~ `Shola Grassland` + `Shola Forest` + Settlements +  
    `Plantations (Tea & Timber)` + `Agricultural Land` + (1 |  
    scientific_name) + (1 | historical_site_code)
Data: glmm_persProb_grassland

     AIC      BIC   logLik deviance df.resid 
     -60      -39       39      -78       68 

Random effects:

Conditional model:
 Groups               Name        Variance Std.Dev.
 scientific_name      (Intercept) 0.2330   0.4827  
 historical_site_code (Intercept) 0.3089   0.5558  
Number of obs: 77, groups:  scientific_name, 26; historical_site_code, 22

Dispersion parameter for beta family (): 11.3 

Conditional model:
                             Estimate Std. Error z value Pr(>|z|)  
(Intercept)                   -4.2532     1.9406  -2.192   0.0284 *
`Shola Grassland`             -2.7095     1.8651  -1.453   0.1463  
`Shola Forest`                 0.2614     0.2988   0.875   0.3818  
Settlements                    0.5983     0.3089   1.937   0.0528 .
`Plantations (Tea & Timber)`   0.6291     0.5290   1.189   0.2344  
`Agricultural Land`            0.2434     0.4306   0.565   0.5718  
---
Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1

```{r}
## forest birds
glmm_persProb_forest <-  glmm_persProb_habitat %>%
  filter(Habitat.type == "Forest")

glmm_forest <- glmmTMB(persProbChange ~ `Shola Grassland` + `Shola Forest` + `Settlements`+ `Plantations (Tea & Timber)` + `Agricultural Land` + (1|scientific_name) + (1|historical_site_code), 
                        data = glmm_persProb_forest, family = beta_family(link = "logit"))

summary(glmm_forest)
```
Family: beta  ( logit )
Formula:          
persProbChange ~ `Shola Grassland` + `Shola Forest` + Settlements +  
    `Plantations (Tea & Timber)` + `Agricultural Land` + (1 |  
    scientific_name) + (1 | historical_site_code)
Data: glmm_persProb_forest

     AIC      BIC   logLik deviance df.resid 
  -387.6   -356.2    202.8   -405.6      233 

Random effects:

Conditional model:
 Groups               Name        Variance Std.Dev.
 scientific_name      (Intercept) 0.44272  0.6654  
 historical_site_code (Intercept) 0.09844  0.3137  
Number of obs: 242, groups:  scientific_name, 41; historical_site_code, 26

Dispersion parameter for beta family (): 10.7 

Conditional model:
                             Estimate Std. Error z value Pr(>|z|)   
(Intercept)                  -1.55241    1.92961  -0.804   0.4211   
`Shola Grassland`             0.24208    1.90338   0.127   0.8988   
`Shola Forest`               -0.02058    0.16216  -0.127   0.8990   
Settlements                   0.46944    0.16172   2.903   0.0037 **
`Plantations (Tea & Timber)`  0.25003    0.26676   0.937   0.3486   
`Agricultural Land`           0.31640    0.22028   1.436   0.1509   
---
Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1

```{r}
## wooded habitat birds
glmm_persProb_wooded <-  glmm_persProb_habitat %>%
  filter(Habitat.type == "Wooded Habitat")

glmm_wooded <- glmmTMB(persProbChange ~ `Shola Grassland` + `Shola Forest` + `Settlements`+ `Plantations (Tea & Timber)` + `Agricultural Land` + (1|scientific_name) + (1|historical_site_code), 
                        data = glmm_persProb_wooded, family = beta_family(link = "logit"))

summary(glmm_wooded)
```
Family: beta  ( logit )
Formula:          
persProbChange ~ `Shola Grassland` + `Shola Forest` + Settlements +  
    `Plantations (Tea & Timber)` + `Agricultural Land` + (1 |  
    scientific_name) + (1 | historical_site_code)
Data: glmm_persProb_wooded

     AIC      BIC   logLik deviance df.resid 
  -971.0   -933.7    494.5   -989.0      458 

Random effects:

Conditional model:
 Groups               Name        Variance Std.Dev.
 scientific_name      (Intercept) 0.71701  0.8468  
 historical_site_code (Intercept) 0.04037  0.2009  
Number of obs: 467, groups:  scientific_name, 72; historical_site_code, 26

Dispersion parameter for beta family (): 13.4 

Conditional model:
                             Estimate Std. Error z value Pr(>|z|)   
(Intercept)                  -1.30397    0.75446  -1.728  0.08392 . 
`Shola Grassland`             0.05310    0.72540   0.073  0.94164   
`Shola Forest`               -0.09691    0.09989  -0.970  0.33195   
Settlements                   0.32177    0.10476   3.071  0.00213 **
`Plantations (Tea & Timber)` -0.04214    0.16733  -0.252  0.80116   
`Agricultural Land`           0.13151    0.14896   0.883  0.37733   
---
Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1

```{r}
## generalist birds
glmm_persProb_generalist <-  glmm_persProb_habitat %>%
  filter(Habitat.type == "Generalist")

glmm_generalist <- glmmTMB(persProbChange ~ `Shola Grassland` + `Shola Forest` + `Settlements`+ `Plantations (Tea & Timber)` + `Agricultural Land` + (1|scientific_name) + (1|historical_site_code), 
                        data = glmm_persProb_generalist, family = beta_family(link = "logit"))

summary(glmm_generalist)
```
Family: beta  ( logit )
Formula:          
persProbChange ~ `Shola Grassland` + `Shola Forest` + Settlements +  
    `Plantations (Tea & Timber)` + `Agricultural Land` + (1 |  
    scientific_name) + (1 | historical_site_code)
Data: glmm_persProb_generalist

     AIC      BIC   logLik deviance df.resid 
  -863.8   -828.9    440.9   -881.8      349 

Random effects:

Conditional model:
 Groups               Name        Variance Std.Dev.
 scientific_name      (Intercept) 0.74008  0.8603  
 historical_site_code (Intercept) 0.05585  0.2363  
Number of obs: 358, groups:  scientific_name, 48; historical_site_code, 26

Dispersion parameter for beta family (): 16.6 

Conditional model:
                             Estimate Std. Error z value Pr(>|z|)  
(Intercept)                  -2.11597    0.88724  -2.385   0.0171 *
`Shola Grassland`            -0.48478    0.85434  -0.567   0.5704  
`Shola Forest`                0.05631    0.11285   0.499   0.6178  
Settlements                   0.27041    0.11955   2.262   0.0237 *
`Plantations (Tea & Timber)`  0.05235    0.19219   0.272   0.7853  
`Agricultural Land`          -0.27848    0.17227  -1.617   0.1060  
---
Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1


## Run mixed-effects models examining species relative abundance and land cover change
```{r}
names(relAbun) <- c("scientific_name","historical_site_code","year","relAbundance")

pivot_1850 <- relAbun %>%
  filter(year != 1900) %>%
  filter(year != 2021) %>%
  pivot_wider(., names_from = "year", values_from = "relAbundance")

pivot_2021 <- relAbun %>%
  filter(year != 1900) %>%
  filter(year != 1850) %>%
  pivot_wider(., names_from = "year", values_from = "relAbundance")

relAbun_wide <- full_join(pivot_1850, pivot_2021, by=c("scientific_name","historical_site_code")) %>%
  replace(is.na(.), 0) %>%
  mutate("relAbunChange" = `1850` - `2021`)

glmm_relAbun <- left_join(lc_change_glmm, relAbun_wide, by="historical_site_code") %>% as.data.frame()

## run the glmm
glmm_lc_relAbun <- glmer(relAbunChange ~ `Shola Grassland` + `Shola Forest` + `Settlements`+ `Plantations (Tea & Timber)` + `Agricultural Land` + (1|scientific_name) + (1|historical_site_code), 
                        data = glmm_relAbun, family = gaussian(link="identity"))

summary(glmm_lc_relAbun)
plot_model(glmm_lc_relAbun, type="pred")
report::report(glmm_lc_relAbun)

# We fitted a linear mixed model (estimated using REML and nloptwrap optimizer) to predict relAbunChange with Shola Grassland, Shola Forest, Settlements, Plantations (Tea & Timber) and Agricultural Land (formula: relAbunChange ~ `Shola Grassland` + `Shola Forest` + Settlements + `Plantations (Tea & Timber)` + `Agricultural Land`). The model included scientific_name and historical_site_code as random effects (formula: list(~1 | scientific_name, ~1 | historical_site_code)). The model's total explanatory power is moderate (conditional R2 = 0.13) and the part related to the fixed effects alone (marginal R2) is of 2.07e-03. The model's intercept, corresponding to Shola Grassland = 0, Shola Forest = 0, Settlements = 0, Plantations (Tea & Timber) = 0 and Agricultural Land = 0, is at -8.92e-03 (95% CI [-0.02, 1.72e-03], t(4905) = -1.64, p = 0.100). Within this model:

#  - The effect of Shola Grassland is statistically non-significant and negative (beta = -3.37e-03, 95% CI [-0.01, 6.85e-03], t(4905) = -0.65, p = 0.518; Std. beta = -0.01, 95% CI [-0.04, 0.02])
 # - The effect of Shola Forest is statistically non-significant and negative (beta = -7.56e-04, 95% CI [-2.44e-03, 9.28e-04], t(4905) = -0.88, p = 0.379; Std. beta = -0.02, 95% CI [-0.05, 0.02])
#  - The effect of Settlements is statistically non-significant and positive (beta = 9.92e-04, 95% CI [-7.97e-04, 2.78e-03], t(4905) = 1.09, p = 0.277; Std. beta = 0.02, 95% CI [-0.01, 0.05])
#  - The effect of Plantations (Tea & Timber) is statistically non-significant and positive (beta = 2.32e-03, 95% CI [-4.72e-04, 5.11e-03], t(4905) = 1.63, p = 0.103; Std. beta = 0.03, 95% CI [-5.35e-03, 0.06])
#  - The effect of Agricultural Land is statistically non-significant and positive (beta = 1.20e-03, 95% CI [-1.35e-03, 3.74e-03], t(4905) = 0.92, p = 0.357; Std. beta = 0.01, 95% CI [-0.02, 0.05])

# Standardized parameters were obtained by fitting the model on a standardized version of the dataset. 95% Confidence Intervals (CIs) and p-values were computed using the Wald approximation
```


## Run mixed-effects models examining persistence probabilties and climate change
```{r}
## we have four predictors for climate
## prep data at the level of the historical site
clim_change_glmm <- clim_change_dat %>%
  dplyr::select(historical_site_code, t_mean_dry_2010, t_mean_rainy_2010, ppt_dry_2010, ppt_rainy_2010) %>%
  group_by(historical_site_code) %>%
  summarise(across(everything(), mean))

## dataframe for glmm
glmm_cc_change <- left_join(persProb_change, clim_change_glmm, by="historical_site_code") %>% as.data.frame()

## scale the climate data
glmm_cc_change <- glmm_cc_change %>%
  mutate(t_mean_dry_2010_z = scale_z(t_mean_dry_2010)) %>%
  mutate(t_mean_rainy_2010_z = scale_z(t_mean_rainy_2010)) %>%
  mutate(ppt_dry_2010_z = scale_z(ppt_dry_2010)) %>%
  mutate(ppt_rainy_2010_z = scale_z(ppt_rainy_2010)) %>%
  as.data.frame()

## run the glmm
glmm_cc <- glmer(persProbChange ~ t_mean_dry_2010_z + t_mean_rainy_2010_z +
                   ppt_dry_2010_z + ppt_rainy_2010_z +
                          (1|scientific_name) + (1|historical_site_code), 
                        data = glmm_cc_change, family = gaussian(link="identity"))

summary(glmm_cc)
plot_model(glmm_cc, type="pred")
plot_model(glmm_cc, type = "est")
report::report(glmm_cc)

# We fitted a linear mixed model (estimated using REML and nloptwrap optimizer) to predict persProbChange with t_mean_dry_2010_z, t_mean_rainy_2010_z, ppt_dry_2010_z and ppt_rainy_2010_z (formula: persProbChange ~ t_mean_dry_2010_z + t_mean_rainy_2010_z + ppt_dry_2010_z + ppt_rainy_2010_z). The model included scientific_name and historical_site_code as random effects (formula: list(~1 | scientific_name, ~1 | historical_site_code)). The model's total explanatory power is substantial (conditional R2 = 0.60) and the part related to the fixed effects alone (marginal R2) is of 9.72e-03. The model's intercept, corresponding to t_mean_dry_2010_z = 0, t_mean_rainy_2010_z = 0, ppt_dry_2010_z = 0 and ppt_rainy_2010_z = 0, is at 0.22 (95% CI [0.19, 0.24], t(1250) = 15.48, p < .001). Within this model:

#  - The effect of t mean dry 2010 z is statistically non-significant and positive (beta = 0.02, 95% CI [-0.15, 0.20], t(1250) = 0.26, p = 0.795; Std. beta = 0.14, 95% CI [-0.90, 1.18])
#  - The effect of t mean rainy 2010 z is statistically non-significant and negative (beta = -0.02, 95% CI [-0.08, 0.04], t(1250) = -0.72, p = 0.472; Std. beta = -0.13, 95% CI [-0.50, 0.23])
#  - The effect of ppt dry 2010 z is statistically non-significant and negative (beta = -0.02, 95% CI [-0.04, 0.01], t(1250) = -1.11, p = 0.268; Std. beta = -0.09, 95% CI [-0.26, 0.07])
#  - The effect of ppt rainy 2010 z is statistically non-significant and negative (beta = -8.50e-03, 95% CI [-0.14, 0.12], t(1250) = -0.13, p = 0.895; Std. beta = -0.05, 95% CI [-0.80, 0.70])

# Standardized parameters were obtained by fitting the model on a standardized version of the dataset. 95% Confidence Intervals (CIs) and p-values were computed using the Wald approximation. 
```
